adapter: Split up catalog.rs
This commit creates two new submodules in the `mz_adapter::catalog`
module by extracting code out of `catalog.rs`. Those two submodules
are:

  - `mz_adapter::catalog::open`: Logic responsible for opening a
  Catalog.
  - `mz_adapter::catalog::objects`: Types used to represent objects
  within a Catalog.

This shortens the length of catalog.rs which has grown to an
unmanageable size.

All the changes in this PR are code movement, visibility modifications,
and switching from direct field access to accessor methods. There is no
change to any logic.
