+++ src/adapter/src/catalog.rs
@@ -10,53 +10,38 @@
 //! Persistent metadata storage for the coordinator.
 
 use std::borrow::Cow;
-use std::collections::{BTreeMap, BTreeSet, VecDeque};
-use std::ops::{Deref, DerefMut};
-use std::str::FromStr;
+use std::collections::{BTreeMap, BTreeSet};
 use std::sync::{atomic, Arc};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use chrono::{DateTime, Utc};
 use futures::Future;
 use itertools::Itertools;
-use once_cell::sync::Lazy;
-use regex::Regex;
-use serde::ser::SerializeSeq;
-use serde::{Deserialize, Serialize};
 use tokio::sync::mpsc::{self, UnboundedSender};
 use tokio::sync::MutexGuard;
-use tracing::{info, trace, warn};
-use uuid::Uuid;
+use tracing::{info, trace};
 
 use mz_audit_log::{EventDetails, EventType, FullNameV1, IdFullNameV1, ObjectType, VersionedEvent};
 use mz_build_info::DUMMY_BUILD_INFO;
 use mz_catalog::builtin::{
-    Builtin, BuiltinCluster, BuiltinLog, BuiltinSource, BuiltinTable, BuiltinType, Fingerprint,
-    BUILTINS, BUILTIN_CLUSTERS, BUILTIN_PREFIXES, BUILTIN_ROLES, MZ_INTROSPECTION_CLUSTER,
-    MZ_SYSTEM_CLUSTER,
+    BuiltinCluster, BuiltinLog, BuiltinSource, BuiltinTable, BuiltinType, BUILTINS,
+    BUILTIN_PREFIXES, MZ_INTROSPECTION_CLUSTER,
 };
-use mz_catalog::objects::{SystemObjectDescription, SystemObjectUniqueIdentifier};
 use mz_catalog::{
-    BootstrapArgs, DurableCatalogState, OpenableDurableCatalogState, StashConfig,
-    SystemObjectMapping, Transaction,
+    BootstrapArgs, DurableCatalogState, OpenableDurableCatalogState, StashConfig, Transaction,
 };
-use mz_compute_client::controller::ComputeReplicaConfig;
-use mz_compute_client::logging::LogVariant;
 use mz_compute_types::dataflows::DataflowDescription;
 use mz_controller::clusters::{
-    ClusterEvent, ClusterRole, ClusterStatus, ManagedReplicaAvailabilityZones,
-    ManagedReplicaLocation, ProcessId, ReplicaAllocation, ReplicaConfig, ReplicaLocation,
-    ReplicaLogging, UnmanagedReplicaLocation,
+    ClusterEvent, ManagedReplicaAvailabilityZones, ManagedReplicaLocation, ReplicaAllocation,
+    ReplicaConfig, ReplicaLocation, UnmanagedReplicaLocation,
 };
 use mz_controller_types::{ClusterId, ReplicaId};
-use mz_expr::{MirScalarExpr, OptimizedMirRelationExpr};
+use mz_expr::OptimizedMirRelationExpr;
 use mz_ore::cast::CastFrom;
 use mz_ore::collections::CollectionExt;
 use mz_ore::metrics::MetricsRegistry;
-use mz_ore::now::{to_datetime, EpochMillis, NowFn};
+use mz_ore::now::{EpochMillis, NowFn};
 use mz_ore::option::FallibleMapExt;
 use mz_ore::result::ResultExt as _;
-use mz_pgrepr::oid::FIRST_USER_OID;
 use mz_repr::adt::mz_acl_item::{merge_mz_acl_items, AclMode, MzAclItem, PrivilegeMap};
 use mz_repr::explain::ExprHumanizer;
 use mz_repr::namespaces::MZ_TEMP_SCHEMA;
@@ -64,7 +49,6 @@ use mz_repr::role_id::RoleId;
 use mz_repr::{Diff, GlobalId, RelationDesc, ScalarType};
 use mz_secrets::InMemorySecretsController;
 use mz_sql::ast::display::AstDisplay;
-use mz_sql::ast::Expr;
 use mz_sql::catalog::{
     CatalogCluster, CatalogClusterReplica, CatalogDatabase, CatalogError as SqlCatalogError,
     CatalogItem as SqlCatalogItem, CatalogItemType as SqlCatalogItemType, CatalogItemType,
@@ -72,48 +56,43 @@ use mz_sql::catalog::{
     DefaultPrivilegeAclItem, DefaultPrivilegeObject, EnvironmentId, IdReference, NameReference,
     RoleAttributes, RoleMembership, RoleVars, SessionCatalog, SystemObjectType,
 };
-use mz_sql::func::OP_IMPLS;
 use mz_sql::names::{
-    Aug, CommentObjectId, DatabaseId, FullItemName, FullSchemaName, ItemQualifiers, ObjectId,
-    PartialItemName, QualifiedItemName, QualifiedSchemaName, ResolvedDatabaseSpecifier,
-    ResolvedIds, SchemaId, SchemaSpecifier, SystemObjectId, PUBLIC_ROLE_NAME,
+    CommentObjectId, DatabaseId, FullItemName, FullSchemaName, ItemQualifiers, ObjectId,
+    PartialItemName, QualifiedItemName, QualifiedSchemaName, ResolvedDatabaseSpecifier, SchemaId,
+    SchemaSpecifier, SystemObjectId, PUBLIC_ROLE_NAME,
 };
 use mz_sql::plan::{
     CreateConnectionPlan, CreateIndexPlan, CreateMaterializedViewPlan, CreateSecretPlan,
-    CreateSinkPlan, CreateSourcePlan, CreateTablePlan, CreateTypePlan, CreateViewPlan,
-    Ingestion as PlanIngestion, Params, Plan, PlanContext, PlanNotice,
-    SourceSinkClusterConfig as PlanStorageClusterConfig, StatementDesc, WebhookHeaders,
-    WebhookValidation,
+    CreateSinkPlan, CreateSourcePlan, CreateTablePlan, CreateTypePlan, CreateViewPlan, Params,
+    Plan, PlanContext, PlanNotice, SourceSinkClusterConfig as PlanStorageClusterConfig,
+    StatementDesc,
 };
 use mz_sql::session::user::{MZ_SYSTEM_ROLE_ID, SUPPORT_USER, SYSTEM_USER};
-use mz_sql::session::vars::{
-    ConnectionCounter, OwnedVarInput, SystemVars, Var, VarError, VarInput, CONFIG_HAS_SYNCED_ONCE,
-};
+use mz_sql::session::vars::{ConnectionCounter, OwnedVarInput, SystemVars};
 use mz_sql::{plan, rbac, DEFAULT_SCHEMA};
 use mz_sql_parser::ast::{
     CreateSinkOption, CreateSourceOption, QualifiedReplica, Statement, WithOptionValue,
 };
-use mz_ssh_util::keys::SshKeyPairSet;
 use mz_stash::{DebugStashFactory, StashFactory};
-use mz_storage_client::controller::IntrospectionType;
-use mz_storage_types::connections::inline::{
-    ConnectionResolver, InlinedConnection, ReferencedConnection,
-};
-use mz_storage_types::sinks::{SinkEnvelope, StorageSinkConnection, StorageSinkConnectionBuilder};
-use mz_storage_types::sources::{
-    IngestionDescription, SourceConnection, SourceDesc, SourceEnvelope, SourceExport, Timeline,
-};
+use mz_storage_types::connections::inline::{ConnectionResolver, InlinedConnection};
+use mz_storage_types::sources::Timeline;
 use mz_transform::dataflow::DataflowMetainfo;
 use mz_transform::Optimizer;
 
 pub use crate::catalog::builtin_table_updates::BuiltinTableUpdate;
 pub use crate::catalog::config::{AwsPrincipalContext, ClusterReplicaSizeMap, Config};
 pub use crate::catalog::error::{AmbiguousRename, Error, ErrorKind};
+pub use crate::catalog::objects::{
+    CatalogEntry, CatalogItem, Cluster, ClusterConfig, ClusterReplica, ClusterReplicaProcessStatus,
+    ClusterVariant, ClusterVariantManaged, CommentsMap, Connection, DataSourceDesc, Database,
+    DefaultPrivileges, Func, Index, Log, MaterializedView, Role, Schema, Secret, Sink, Source,
+    StorageSinkConnectionState, Table, Type, View,
+};
+pub use crate::catalog::open::BuiltinMigrationMetadata;
 pub use crate::catalog::state::CatalogState;
 use crate::client::ConnectionId;
 use crate::command::CatalogDump;
-use crate::config::{SynchronizedParameters, SystemParameterFrontend, SystemParameterSyncConfig};
-use crate::coord::{timestamp_oracle, ConnMeta, TargetCluster, DEFAULT_LOGICAL_COMPACTION_WINDOW};
+use crate::coord::{ConnMeta, TargetCluster};
 use crate::session::{PreparedStatement, Session, DEFAULT_DATABASE_NAME};
 use crate::util::ResultExt;
 use crate::{AdapterError, AdapterNotice, ExecuteResponse};
@@ -125,6 +104,8 @@ mod error;
 mod migrate;
 
 mod inner;
+mod objects;
+mod open;
 mod state;
 
 pub static SYSTEM_CONN_ID: ConnectionId = ConnectionId::Static(0);
@@ -341,2681 +322,13 @@ impl ConnectionResolver for ConnCatalog<'_> {
     }
 }
 
-#[derive(Debug, Serialize, Clone)]
-pub struct Database {
-    pub name: String,
-    pub id: DatabaseId,
-    #[serde(skip)]
-    pub oid: u32,
-    #[serde(serialize_with = "mz_ore::serde::map_key_to_string")]
-    pub schemas_by_id: BTreeMap<SchemaId, Schema>,
-    pub schemas_by_name: BTreeMap<String, SchemaId>,
-    pub owner_id: RoleId,
-    pub privileges: PrivilegeMap,
-}
-
-impl From<Database> for mz_catalog::Database {
-    fn from(database: Database) -> mz_catalog::Database {
-        mz_catalog::Database {
-            id: database.id,
-            name: database.name,
-            owner_id: database.owner_id,
-            privileges: database.privileges.into_all_values().collect(),
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct Schema {
-    pub name: QualifiedSchemaName,
-    pub id: SchemaSpecifier,
-    #[serde(skip)]
-    pub oid: u32,
-    pub items: BTreeMap<String, GlobalId>,
-    pub functions: BTreeMap<String, GlobalId>,
-    pub owner_id: RoleId,
-    pub privileges: PrivilegeMap,
-}
-
-impl Schema {
-    fn into_durable_schema(self, database_id: Option<DatabaseId>) -> mz_catalog::Schema {
-        mz_catalog::Schema {
-            id: self.id.into(),
-            name: self.name.schema,
-            database_id,
-            owner_id: self.owner_id,
-            privileges: self.privileges.into_all_values().collect(),
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct Role {
-    pub name: String,
-    pub id: RoleId,
-    #[serde(skip)]
-    pub oid: u32,
-    pub attributes: RoleAttributes,
-    pub membership: RoleMembership,
-    pub vars: RoleVars,
-}
-
-impl Role {
-    pub fn is_user(&self) -> bool {
-        self.id.is_user()
-    }
-
-    pub fn vars<'a>(&'a self) -> impl Iterator<Item = (&'a str, &'a OwnedVarInput)> {
-        self.vars.map.iter().map(|(name, val)| (name.as_str(), val))
-    }
-}
-
-impl From<Role> for mz_catalog::Role {
-    fn from(role: Role) -> mz_catalog::Role {
-        mz_catalog::Role {
-            id: role.id,
-            name: role.name,
-            attributes: role.attributes,
-            membership: role.membership,
-            vars: role.vars,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct Cluster {
-    pub name: String,
-    pub id: ClusterId,
-    pub config: ClusterConfig,
-    #[serde(skip)]
-    pub log_indexes: BTreeMap<LogVariant, GlobalId>,
-    pub linked_object_id: Option<GlobalId>,
-    /// Objects bound to this cluster. Does not include introspection source
-    /// indexes.
-    pub bound_objects: BTreeSet<GlobalId>,
-    replica_id_by_name_: BTreeMap<String, ReplicaId>,
-    #[serde(serialize_with = "mz_ore::serde::map_key_to_string")]
-    replicas_by_id_: BTreeMap<ReplicaId, ClusterReplica>,
-    pub owner_id: RoleId,
-    pub privileges: PrivilegeMap,
-}
-
-impl Cluster {
-    /// The role of the cluster. Currently used to set alert severity.
-    pub fn role(&self) -> ClusterRole {
-        // NOTE - These roles power monitoring systems. Do not change
-        // them without talking to the cloud or observability groups.
-        if self.name == MZ_SYSTEM_CLUSTER.name {
-            ClusterRole::SystemCritical
-        } else if self.name == MZ_INTROSPECTION_CLUSTER.name {
-            ClusterRole::System
-        } else {
-            ClusterRole::User
-        }
-    }
-
-    /// Returns `true` if the cluster is a managed cluster.
-    pub fn is_managed(&self) -> bool {
-        matches!(self.config.variant, ClusterVariant::Managed { .. })
-    }
-
-    /// Lists the user replicas, which are those that do not have the internal flag set.
-    pub fn user_replicas(&self) -> impl Iterator<Item = &ClusterReplica> {
-        self.replicas().filter(|r| !r.config.location.internal())
-    }
-
-    /// Lists all replicas in the cluster
-    pub fn replicas(&self) -> impl Iterator<Item = &ClusterReplica> {
-        self.replicas_by_id_.values()
-    }
-
-    /// Lookup a replica by ID.
-    pub fn replica(&self, replica_id: ReplicaId) -> Option<&ClusterReplica> {
-        self.replicas_by_id_.get(&replica_id)
-    }
-
-    /// Lookup a replica by ID and return a mutable reference.
-    fn replica_mut(&mut self, replica_id: ReplicaId) -> Option<&mut ClusterReplica> {
-        self.replicas_by_id_.get_mut(&replica_id)
-    }
-
-    /// Lookup a replica ID by name.
-    pub fn replica_id(&self, name: &str) -> Option<ReplicaId> {
-        self.replica_id_by_name_.get(name).copied()
-    }
-
-    /// Insert a new replica into the cluster.
-    ///
-    /// Panics if the name or ID are reused.
-    fn insert_replica(&mut self, replica: ClusterReplica) {
-        assert!(self
-            .replica_id_by_name_
-            .insert(replica.name.clone(), replica.replica_id)
-            .is_none());
-        assert!(self
-            .replicas_by_id_
-            .insert(replica.replica_id, replica)
-            .is_none());
-    }
-
-    /// Remove a replica from this cluster.
-    ///
-    /// Panics if the replica ID does not exist, or if the internal state is inconsistent.
-    fn remove_replica(&mut self, replica_id: ReplicaId) {
-        let replica = self
-            .replicas_by_id_
-            .remove(&replica_id)
-            .expect("catalog out of sync");
-        self.replica_id_by_name_
-            .remove(&replica.name)
-            .expect("catalog out of sync");
-        assert_eq!(self.replica_id_by_name_.len(), self.replicas_by_id_.len());
-    }
-
-    /// Renames a replica to a new name.
-    ///
-    /// Panics if the replica ID is unknown, or new name is not unique, or the internal state is
-    /// inconsistent.
-    fn rename_replica(&mut self, replica_id: ReplicaId, to_name: String) {
-        let replica = self.replica_mut(replica_id).expect("Must exist");
-        let old_name = std::mem::take(&mut replica.name);
-        replica.name = to_name.clone();
-
-        assert!(self.replica_id_by_name_.remove(&old_name).is_some());
-        assert!(self
-            .replica_id_by_name_
-            .insert(to_name, replica_id)
-            .is_none());
-    }
-}
-
-impl From<Cluster> for mz_catalog::Cluster {
-    fn from(cluster: Cluster) -> mz_catalog::Cluster {
-        mz_catalog::Cluster {
-            id: cluster.id,
-            name: cluster.name,
-            linked_object_id: cluster.linked_object_id,
-            owner_id: cluster.owner_id,
-            privileges: cluster.privileges.into_all_values().collect(),
-            config: cluster.config.into(),
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct ClusterReplica {
-    pub name: String,
-    pub cluster_id: ClusterId,
-    pub replica_id: ReplicaId,
-    pub config: ReplicaConfig,
-    #[serde(skip)]
-    pub process_status: BTreeMap<ProcessId, ClusterReplicaProcessStatus>,
-    pub owner_id: RoleId,
-}
-
-impl ClusterReplica {
-    /// Computes the status of the cluster replica as a whole.
-    pub fn status(&self) -> ClusterStatus {
-        self.process_status
-            .values()
-            .fold(ClusterStatus::Ready, |s, p| match (s, p.status) {
-                (ClusterStatus::Ready, ClusterStatus::Ready) => ClusterStatus::Ready,
-                (x, y) => {
-                    let reason_x = match x {
-                        ClusterStatus::NotReady(reason) => reason,
-                        ClusterStatus::Ready => None,
-                    };
-                    let reason_y = match y {
-                        ClusterStatus::NotReady(reason) => reason,
-                        ClusterStatus::Ready => None,
-                    };
-                    // Arbitrarily pick the first known not-ready reason.
-                    ClusterStatus::NotReady(reason_x.or(reason_y))
-                }
-            })
-    }
-}
-
-impl From<ClusterReplica> for mz_catalog::ClusterReplica {
-    fn from(replica: ClusterReplica) -> mz_catalog::ClusterReplica {
-        mz_catalog::ClusterReplica {
-            cluster_id: replica.cluster_id,
-            replica_id: replica.replica_id,
-            name: replica.name,
-            config: replica.config.into(),
-            owner_id: replica.owner_id,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Clone)]
-pub struct ClusterReplicaProcessStatus {
-    pub status: ClusterStatus,
-    pub time: DateTime<Utc>,
-}
-
-#[derive(Clone, Debug, Serialize)]
-pub struct CatalogEntry {
-    item: CatalogItem,
-    #[serde(skip)]
-    used_by: Vec<GlobalId>,
-    id: GlobalId,
-    #[serde(skip)]
-    oid: u32,
-    name: QualifiedItemName,
-    owner_id: RoleId,
-    privileges: PrivilegeMap,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub enum CatalogItem {
-    Table(Table),
-    Source(Source),
-    Log(Log),
-    View(View),
-    MaterializedView(MaterializedView),
-    Sink(Sink),
-    Index(Index),
-    Type(Type),
-    Func(Func),
-    Secret(Secret),
-    Connection(Connection),
-}
-
-impl From<CatalogEntry> for mz_catalog::Item {
-    fn from(entry: CatalogEntry) -> mz_catalog::Item {
-        mz_catalog::Item {
-            id: entry.id,
-            name: entry.name,
-            create_sql: entry.item.into_serialized(),
-            owner_id: entry.owner_id,
-            privileges: entry.privileges.into_all_values().collect(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Table {
-    pub create_sql: String,
-    pub desc: RelationDesc,
-    #[serde(skip)]
-    pub defaults: Vec<Expr<Aug>>,
-    #[serde(skip)]
-    pub conn_id: Option<ConnectionId>,
-    pub resolved_ids: ResolvedIds,
-    pub custom_logical_compaction_window: Option<Duration>,
-    /// Whether the table's logical compaction window is controlled by
-    /// METRICS_RETENTION
-    pub is_retained_metrics_object: bool,
-}
-
-impl Table {
-    // The Coordinator controls insertions for tables (including system tables),
-    // so they are realtime.
-    pub fn timeline(&self) -> Timeline {
-        Timeline::EpochMilliseconds
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub enum DataSourceDesc {
-    /// Receives data from an external system
-    Ingestion(IngestionDescription<(), ReferencedConnection>),
-    /// Receives data from some other source
-    Source,
-    /// Receives introspection data from an internal system
-    Introspection(IntrospectionType),
-    /// Receives data from the source's reclocking/remapping operations.
-    Progress,
-    /// Receives data from HTTP requests.
-    Webhook {
-        /// Optional components used to validation a webhook request.
-        validate_using: Option<WebhookValidation>,
-        /// Describes whether or not to include headers and how to map them.
-        headers: WebhookHeaders,
-        /// The cluster which this source is associated with.
-        cluster_id: ClusterId,
-    },
-}
-
-impl DataSourceDesc {
-    /// Describes the ingestion to the adapter, which essentially just enriches the a higher-level
-    /// [`PlanIngestion`] with a [`ClusterId`].
-    pub fn ingestion(
-        id: GlobalId,
-        ingestion: PlanIngestion,
-        instance_id: ClusterId,
-    ) -> DataSourceDesc {
-        let source_imports = ingestion
-            .source_imports
-            .iter()
-            .map(|id| (*id, ()))
-            .collect();
-
-        let source_exports = ingestion
-            .subsource_exports
-            .iter()
-            // By convention the first output corresponds to the main source object
-            .chain(std::iter::once((&id, &0)))
-            .map(|(id, output_index)| {
-                let export = SourceExport {
-                    output_index: *output_index,
-                    storage_metadata: (),
-                };
-                (*id, export)
-            })
-            .collect();
-
-        DataSourceDesc::Ingestion(IngestionDescription {
-            desc: ingestion.desc.clone(),
-            ingestion_metadata: (),
-            source_imports,
-            source_exports,
-            instance_id,
-            remap_collection_id: ingestion.progress_subsource,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Source {
-    pub create_sql: String,
-    // TODO: Unskip: currently blocked on some inner BTreeMap<X, _> problems.
-    #[serde(skip)]
-    pub data_source: DataSourceDesc,
-    pub desc: RelationDesc,
-    pub timeline: Timeline,
-    pub resolved_ids: ResolvedIds,
-    pub custom_logical_compaction_window: Option<Duration>,
-    /// Whether the source's logical compaction window is controlled by
-    /// METRICS_RETENTION
-    pub is_retained_metrics_object: bool,
-}
-
-impl Source {
-    /// Creates a new `Source`.
-    ///
-    /// # Panics
-    /// - If an ingestion-based plan is not given a cluster_id.
-    /// - If a non-ingestion-based source has a defined cluster config in its plan.
-    /// - If a non-ingestion-based source is given a cluster_id.
-    pub fn new(
-        id: GlobalId,
-        plan: CreateSourcePlan,
-        cluster_id: Option<ClusterId>,
-        resolved_ids: ResolvedIds,
-        custom_logical_compaction_window: Option<Duration>,
-        is_retained_metrics_object: bool,
-    ) -> Source {
-        Source {
-            create_sql: plan.source.create_sql,
-            data_source: match plan.source.data_source {
-                mz_sql::plan::DataSourceDesc::Ingestion(ingestion) => DataSourceDesc::ingestion(
-                    id,
-                    ingestion,
-                    cluster_id.expect("ingestion-based sources must be given a cluster ID"),
-                ),
-                mz_sql::plan::DataSourceDesc::Progress => {
-                    assert!(
-                        matches!(
-                            plan.cluster_config,
-                            mz_sql::plan::SourceSinkClusterConfig::Undefined
-                        ) && cluster_id.is_none(),
-                        "subsources must not have a host config or cluster_id defined"
-                    );
-                    DataSourceDesc::Progress
-                }
-                mz_sql::plan::DataSourceDesc::Source => {
-                    assert!(
-                        matches!(
-                            plan.cluster_config,
-                            mz_sql::plan::SourceSinkClusterConfig::Undefined
-                        ) && cluster_id.is_none(),
-                        "subsources must not have a host config or cluster_id defined"
-                    );
-                    DataSourceDesc::Source
-                }
-                mz_sql::plan::DataSourceDesc::Webhook {
-                    validate_using,
-                    headers,
-                } => {
-                    assert!(
-                        matches!(
-                            plan.cluster_config,
-                            mz_sql::plan::SourceSinkClusterConfig::Existing { .. }
-                        ) && cluster_id.is_some(),
-                        "webhook sources must be created on an existing cluster"
-                    );
-                    DataSourceDesc::Webhook {
-                        validate_using,
-                        headers,
-                        cluster_id: cluster_id.expect("checked above"),
-                    }
-                }
-            },
-            desc: plan.source.desc,
-            timeline: plan.timeline,
-            resolved_ids,
-            custom_logical_compaction_window,
-            is_retained_metrics_object,
-        }
-    }
-
-    /// Returns whether this source ingests data from an external source.
-    pub fn is_external(&self) -> bool {
-        match self.data_source {
-            DataSourceDesc::Ingestion(_) | DataSourceDesc::Webhook { .. } => true,
-            DataSourceDesc::Introspection(_)
-            | DataSourceDesc::Progress
-            | DataSourceDesc::Source => false,
-        }
-    }
-
-    /// Type of the source.
-    pub fn source_type(&self) -> &str {
-        match &self.data_source {
-            DataSourceDesc::Ingestion(ingestion) => ingestion.desc.connection.name(),
-            DataSourceDesc::Progress => "progress",
-            DataSourceDesc::Source => "subsource",
-            DataSourceDesc::Introspection(_) => "source",
-            DataSourceDesc::Webhook { .. } => "webhook",
-        }
-    }
-
-    /// Envelope of the source.
-    pub fn envelope(&self) -> Option<&str> {
-        // Note how "none"/"append-only" is different from `None`. Source
-        // sources don't have an envelope (internal logs, for example), while
-        // other sources have an envelope that we call the "NONE"-envelope.
-
-        match &self.data_source {
-            // NOTE(aljoscha): We could move the block for ingestsions into
-            // `SourceEnvelope` itself, but that one feels more like an internal
-            // thing and adapter should own how we represent envelopes as a
-            // string? It would not be hard to convince me otherwise, though.
-            DataSourceDesc::Ingestion(ingestion) => match ingestion.desc.envelope() {
-                SourceEnvelope::None(_) => Some("none"),
-                SourceEnvelope::Debezium(_) => {
-                    // NOTE(aljoscha): This is currently not used in production.
-                    // DEBEZIUM sources transparently use `DEBEZIUM UPSERT`.
-                    Some("debezium")
-                }
-                SourceEnvelope::Upsert(upsert_envelope) => match upsert_envelope.style {
-                    mz_storage_types::sources::UpsertStyle::Default(_) => Some("upsert"),
-                    mz_storage_types::sources::UpsertStyle::Debezium { .. } => {
-                        // NOTE(aljoscha): Should we somehow mark that this is
-                        // using upsert internally? See note above about
-                        // DEBEZIUM.
-                        Some("debezium")
-                    }
-                },
-                SourceEnvelope::CdcV2 => {
-                    // TODO(aljoscha): Should we even report this? It's
-                    // currently not exposed.
-                    Some("materialize")
-                }
-            },
-            DataSourceDesc::Introspection(_)
-            | DataSourceDesc::Webhook { .. }
-            | DataSourceDesc::Progress
-            | DataSourceDesc::Source => None,
-        }
-    }
-
-    /// Connection ID of the source, if one exists.
-    pub fn connection_id(&self) -> Option<GlobalId> {
-        match &self.data_source {
-            DataSourceDesc::Ingestion(ingestion) => ingestion.desc.connection.connection_id(),
-            DataSourceDesc::Introspection(_)
-            | DataSourceDesc::Webhook { .. }
-            | DataSourceDesc::Progress
-            | DataSourceDesc::Source => None,
-        }
-    }
-
-    /// The expensive resource that each source consumes is persist shards. To
-    /// prevent abuse, we want to prevent users from creating sources that use an
-    /// unbounded number of persist shards. But we also don't want to count
-    /// persist shards that are mandated by teh system (e.g., the progress
-    /// shard) so that future versions of Materialize can introduce additional
-    /// per-source shards (e.g., a per-source status shard) without impacting
-    /// the limit calculation.
-    pub fn user_controllable_persist_shard_count(&self) -> i64 {
-        match &self.data_source {
-            DataSourceDesc::Ingestion(ingestion) => {
-                // Ingestions with subsources only use persist shards for their
-                // subsources (i.e. not the primary source's persist shard);
-                // those without subsources use 1 (their primary source's
-                // persist shard).
-                std::cmp::max(1, i64::try_from(ingestion.source_exports.len().saturating_sub(1)).expect("fewer than i64::MAX persist shards"))
-            }
-            DataSourceDesc::Webhook { .. } => 1,
-            //  DataSourceDesc::Source represents subsources, which are accounted for in their
-            //  primary source's ingestion.
-            DataSourceDesc::Source
-            // Introspection and progress subsources are not under the user's control, so shouldn't
-            // count toward their quota.
-            | DataSourceDesc::Introspection(_)
-            | DataSourceDesc::Progress => 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Log {
-    pub variant: LogVariant,
-    /// Whether the log is backed by a storage collection.
-    pub has_storage_collection: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Sink {
-    pub create_sql: String,
-    pub from: GlobalId,
-    // TODO(benesch): this field duplicates information that could be derived
-    // from the connection ID. Too hard to fix at the moment.
-    #[serde(skip)]
-    pub connection: StorageSinkConnectionState,
-    pub envelope: SinkEnvelope,
-    pub with_snapshot: bool,
-    pub resolved_ids: ResolvedIds,
-    pub cluster_id: ClusterId,
-}
-
-impl Sink {
-    pub fn sink_type(&self) -> &str {
-        match &self.connection {
-            StorageSinkConnectionState::Pending(pending) => pending.name(),
-            StorageSinkConnectionState::Ready(ready) => ready.name(),
-        }
-    }
-
-    /// Envelope of the sink.
-    pub fn envelope(&self) -> Option<&str> {
-        match &self.envelope {
-            SinkEnvelope::Debezium => Some("debezium"),
-            SinkEnvelope::Upsert => Some("upsert"),
-        }
-    }
-
-    pub fn connection_id(&self) -> Option<GlobalId> {
-        match &self.connection {
-            StorageSinkConnectionState::Pending(pending) => pending.connection_id(),
-            StorageSinkConnectionState::Ready(ready) => ready.connection_id(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub enum StorageSinkConnectionState {
-    Pending(StorageSinkConnectionBuilder<ReferencedConnection>),
-    Ready(StorageSinkConnection<ReferencedConnection>),
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct View {
-    pub create_sql: String,
-    pub optimized_expr: OptimizedMirRelationExpr,
-    pub desc: RelationDesc,
-    pub conn_id: Option<ConnectionId>,
-    pub resolved_ids: ResolvedIds,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct MaterializedView {
-    pub create_sql: String,
-    pub optimized_expr: OptimizedMirRelationExpr,
-    pub desc: RelationDesc,
-    pub resolved_ids: ResolvedIds,
-    pub cluster_id: ClusterId,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Index {
-    pub create_sql: String,
-    pub on: GlobalId,
-    pub keys: Vec<MirScalarExpr>,
-    pub conn_id: Option<ConnectionId>,
-    pub resolved_ids: ResolvedIds,
-    pub cluster_id: ClusterId,
-    pub custom_logical_compaction_window: Option<Duration>,
-    pub is_retained_metrics_object: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Type {
-    pub create_sql: String,
-    #[serde(skip)]
-    pub details: CatalogTypeDetails<IdReference>,
-    pub resolved_ids: ResolvedIds,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Func {
-    #[serde(skip)]
-    pub inner: &'static mz_sql::func::Func,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Secret {
-    pub create_sql: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Connection {
-    pub create_sql: String,
-    pub connection: mz_storage_types::connections::Connection<ReferencedConnection>,
-    pub resolved_ids: ResolvedIds,
-}
-
 pub struct TransactionResult<R> {
     pub builtin_table_updates: Vec<BuiltinTableUpdate>,
     pub audit_events: Vec<VersionedEvent>,
     pub result: R,
 }
 
-impl CatalogItem {
-    /// Returns a string indicating the type of this catalog entry.
-    pub(crate) fn typ(&self) -> mz_sql::catalog::CatalogItemType {
-        match self {
-            CatalogItem::Table(_) => mz_sql::catalog::CatalogItemType::Table,
-            CatalogItem::Source(_) => mz_sql::catalog::CatalogItemType::Source,
-            CatalogItem::Log(_) => mz_sql::catalog::CatalogItemType::Source,
-            CatalogItem::Sink(_) => mz_sql::catalog::CatalogItemType::Sink,
-            CatalogItem::View(_) => mz_sql::catalog::CatalogItemType::View,
-            CatalogItem::MaterializedView(_) => mz_sql::catalog::CatalogItemType::MaterializedView,
-            CatalogItem::Index(_) => mz_sql::catalog::CatalogItemType::Index,
-            CatalogItem::Type(_) => mz_sql::catalog::CatalogItemType::Type,
-            CatalogItem::Func(_) => mz_sql::catalog::CatalogItemType::Func,
-            CatalogItem::Secret(_) => mz_sql::catalog::CatalogItemType::Secret,
-            CatalogItem::Connection(_) => mz_sql::catalog::CatalogItemType::Connection,
-        }
-    }
-
-    pub fn desc(&self, name: &FullItemName) -> Result<Cow<RelationDesc>, SqlCatalogError> {
-        match &self {
-            CatalogItem::Source(src) => Ok(Cow::Borrowed(&src.desc)),
-            CatalogItem::Log(log) => Ok(Cow::Owned(log.variant.desc())),
-            CatalogItem::Table(tbl) => Ok(Cow::Borrowed(&tbl.desc)),
-            CatalogItem::View(view) => Ok(Cow::Borrowed(&view.desc)),
-            CatalogItem::MaterializedView(mview) => Ok(Cow::Borrowed(&mview.desc)),
-            CatalogItem::Func(_)
-            | CatalogItem::Index(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => Err(SqlCatalogError::InvalidDependency {
-                name: name.to_string(),
-                typ: self.typ(),
-            }),
-        }
-    }
-
-    pub fn func(
-        &self,
-        entry: &CatalogEntry,
-    ) -> Result<&'static mz_sql::func::Func, SqlCatalogError> {
-        match &self {
-            CatalogItem::Func(func) => Ok(func.inner),
-            _ => Err(SqlCatalogError::UnexpectedType {
-                name: entry.name().item.to_string(),
-                actual_type: entry.item_type(),
-                expected_type: CatalogItemType::Func,
-            }),
-        }
-    }
-
-    pub fn source_desc(
-        &self,
-        entry: &CatalogEntry,
-    ) -> Result<Option<&SourceDesc<ReferencedConnection>>, SqlCatalogError> {
-        match &self {
-            CatalogItem::Source(source) => match &source.data_source {
-                DataSourceDesc::Ingestion(ingestion) => Ok(Some(&ingestion.desc)),
-                DataSourceDesc::Introspection(_)
-                | DataSourceDesc::Webhook { .. }
-                | DataSourceDesc::Progress
-                | DataSourceDesc::Source => Ok(None),
-            },
-            _ => Err(SqlCatalogError::UnexpectedType {
-                name: entry.name().item.to_string(),
-                actual_type: entry.item_type(),
-                expected_type: CatalogItemType::Source,
-            }),
-        }
-    }
-
-    /// Collects the identifiers of the objects that were encountered when
-    /// resolving names in the item's DDL statement.
-    pub fn uses(&self) -> &ResolvedIds {
-        static EMPTY: Lazy<ResolvedIds> = Lazy::new(|| ResolvedIds(BTreeSet::new()));
-        match self {
-            CatalogItem::Func(_) => &*EMPTY,
-            CatalogItem::Index(idx) => &idx.resolved_ids,
-            CatalogItem::Sink(sink) => &sink.resolved_ids,
-            CatalogItem::Source(source) => &source.resolved_ids,
-            CatalogItem::Log(_) => &*EMPTY,
-            CatalogItem::Table(table) => &table.resolved_ids,
-            CatalogItem::Type(typ) => &typ.resolved_ids,
-            CatalogItem::View(view) => &view.resolved_ids,
-            CatalogItem::MaterializedView(mview) => &mview.resolved_ids,
-            CatalogItem::Secret(_) => &*EMPTY,
-            CatalogItem::Connection(connection) => &connection.resolved_ids,
-        }
-    }
-
-    /// Indicates whether this item is a placeholder for a future item
-    /// or if it's actually a real item.
-    pub fn is_placeholder(&self) -> bool {
-        match self {
-            CatalogItem::Func(_)
-            | CatalogItem::Index(_)
-            | CatalogItem::Source(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::Table(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::View(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => false,
-            CatalogItem::Sink(s) => match s.connection {
-                StorageSinkConnectionState::Pending(_) => true,
-                StorageSinkConnectionState::Ready(_) => false,
-            },
-        }
-    }
-
-    /// Returns the connection ID that this item belongs to, if this item is
-    /// temporary.
-    pub fn conn_id(&self) -> Option<&ConnectionId> {
-        match self {
-            CatalogItem::View(view) => view.conn_id.as_ref(),
-            CatalogItem::Index(index) => index.conn_id.as_ref(),
-            CatalogItem::Table(table) => table.conn_id.as_ref(),
-            CatalogItem::Log(_)
-            | CatalogItem::Source(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Connection(_) => None,
-        }
-    }
-
-    /// Indicates whether this item is temporary or not.
-    pub fn is_temporary(&self) -> bool {
-        self.conn_id().is_some()
-    }
-
-    /// Returns a clone of `self` with all instances of `from` renamed to `to`
-    /// (with the option of including the item's own name) or errors if request
-    /// is ambiguous.
-    fn rename_item_refs(
-        &self,
-        from: FullItemName,
-        to_item_name: String,
-        rename_self: bool,
-    ) -> Result<CatalogItem, String> {
-        let do_rewrite = |create_sql: String| -> Result<String, String> {
-            let mut create_stmt = mz_sql::parse::parse(&create_sql)
-                .expect("invalid create sql persisted to catalog")
-                .into_element()
-                .ast;
-            if rename_self {
-                mz_sql::ast::transform::create_stmt_rename(&mut create_stmt, to_item_name.clone());
-            }
-            // Determination of what constitutes an ambiguous request is done here.
-            mz_sql::ast::transform::create_stmt_rename_refs(&mut create_stmt, from, to_item_name)?;
-            Ok(create_stmt.to_ast_string_stable())
-        };
-
-        match self {
-            CatalogItem::Table(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Table(i))
-            }
-            CatalogItem::Log(i) => Ok(CatalogItem::Log(i.clone())),
-            CatalogItem::Source(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Source(i))
-            }
-            CatalogItem::Sink(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Sink(i))
-            }
-            CatalogItem::View(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::View(i))
-            }
-            CatalogItem::MaterializedView(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::MaterializedView(i))
-            }
-            CatalogItem::Index(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Index(i))
-            }
-            CatalogItem::Secret(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Secret(i))
-            }
-            CatalogItem::Func(_) | CatalogItem::Type(_) => {
-                unreachable!("{}s cannot be renamed", self.typ())
-            }
-            CatalogItem::Connection(i) => {
-                let mut i = i.clone();
-                i.create_sql = do_rewrite(i.create_sql)?;
-                Ok(CatalogItem::Connection(i))
-            }
-        }
-    }
-
-    /// If the object is considered a "compute object"
-    /// (i.e., it is managed by the compute controller),
-    /// this function returns its cluster ID. Otherwise, it returns nothing.
-    ///
-    /// This function differs from `cluster_id` because while all
-    /// compute objects run on a cluster, the converse is not true.
-    pub fn is_compute_object_on_cluster(&self) -> Option<ClusterId> {
-        match self {
-            CatalogItem::Index(index) => Some(index.cluster_id),
-            CatalogItem::Table(_)
-            | CatalogItem::Source(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => None,
-        }
-    }
-
-    pub fn cluster_id(&self) -> Option<ClusterId> {
-        match self {
-            CatalogItem::MaterializedView(mv) => Some(mv.cluster_id),
-            CatalogItem::Index(index) => Some(index.cluster_id),
-            CatalogItem::Source(source) => match &source.data_source {
-                DataSourceDesc::Ingestion(ingestion) => Some(ingestion.instance_id),
-                DataSourceDesc::Webhook { cluster_id, .. } => Some(*cluster_id),
-                DataSourceDesc::Introspection(_)
-                | DataSourceDesc::Progress
-                | DataSourceDesc::Source => None,
-            },
-            CatalogItem::Sink(sink) => Some(sink.cluster_id),
-            CatalogItem::Table(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => None,
-        }
-    }
-
-    /// The custom compaction window, if any has been set.
-    // Note[btv]: As of 2023-04-10, this is only set
-    // for objects with `is_retained_metrics_object`. That
-    // may not always be true in the future, if we enable user-settable
-    // compaction windows.
-    pub fn custom_logical_compaction_window(&self) -> Option<Duration> {
-        match self {
-            CatalogItem::Table(table) => table.custom_logical_compaction_window,
-            CatalogItem::Source(source) => source.custom_logical_compaction_window,
-            CatalogItem::Index(index) => index.custom_logical_compaction_window,
-            CatalogItem::MaterializedView(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => None,
-        }
-    }
-
-    /// The initial compaction window, for objects that have one; that is,
-    /// tables, sources, indexes, and MVs.
-    ///
-    /// If `custom_logical_compaction_window()` returns something, use
-    /// that.  Otherwise, use a sensible default (currently 1s).
-    ///
-    /// For objects that do not have the concept of compaction window,
-    /// return nothing.
-    pub fn initial_logical_compaction_window(&self) -> Option<Duration> {
-        let custom_logical_compaction_window = match self {
-            CatalogItem::Table(_)
-            | CatalogItem::Source(_)
-            | CatalogItem::Index(_)
-            | CatalogItem::MaterializedView(_) => self.custom_logical_compaction_window(),
-            CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => return None,
-        };
-        Some(custom_logical_compaction_window.unwrap_or(DEFAULT_LOGICAL_COMPACTION_WINDOW))
-    }
-
-    /// Whether the item's logical compaction window
-    /// is controlled by the METRICS_RETENTION
-    /// system var.
-    pub fn is_retained_metrics_object(&self) -> bool {
-        match self {
-            CatalogItem::Table(table) => table.is_retained_metrics_object,
-            CatalogItem::Source(source) => source.is_retained_metrics_object,
-            CatalogItem::Index(index) => index.is_retained_metrics_object,
-            CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => false,
-        }
-    }
-
-    pub(crate) fn to_serialized(&self) -> String {
-        match self {
-            CatalogItem::Table(table) => table.create_sql.clone(),
-            CatalogItem::Log(_) => unreachable!("builtin logs cannot be serialized"),
-            CatalogItem::Source(source) => {
-                assert!(
-                    !matches!(source.data_source, DataSourceDesc::Introspection(_)),
-                    "cannot serialize introspection/builtin sources",
-                );
-                source.create_sql.clone()
-            }
-            CatalogItem::View(view) => view.create_sql.clone(),
-            CatalogItem::MaterializedView(mview) => mview.create_sql.clone(),
-            CatalogItem::Index(index) => index.create_sql.clone(),
-            CatalogItem::Sink(sink) => sink.create_sql.clone(),
-            CatalogItem::Type(typ) => typ.create_sql.clone(),
-            CatalogItem::Secret(secret) => secret.create_sql.clone(),
-            CatalogItem::Connection(connection) => connection.create_sql.clone(),
-            CatalogItem::Func(_) => unreachable!("cannot serialize functions yet"),
-        }
-    }
-
-    pub(crate) fn into_serialized(self) -> String {
-        match self {
-            CatalogItem::Table(table) => table.create_sql,
-            CatalogItem::Log(_) => unreachable!("builtin logs cannot be serialized"),
-            CatalogItem::Source(source) => {
-                assert!(
-                    !matches!(source.data_source, DataSourceDesc::Introspection(_)),
-                    "cannot serialize introspection/builtin sources",
-                );
-                source.create_sql
-            }
-            CatalogItem::View(view) => view.create_sql,
-            CatalogItem::MaterializedView(mview) => mview.create_sql,
-            CatalogItem::Index(index) => index.create_sql,
-            CatalogItem::Sink(sink) => sink.create_sql,
-            CatalogItem::Type(typ) => typ.create_sql,
-            CatalogItem::Secret(secret) => secret.create_sql,
-            CatalogItem::Connection(connection) => connection.create_sql,
-            CatalogItem::Func(_) => unreachable!("cannot serialize functions yet"),
-        }
-    }
-}
-
-impl CatalogEntry {
-    /// Reports the description of the datums produced by this catalog item.
-    pub fn desc(&self, name: &FullItemName) -> Result<Cow<RelationDesc>, SqlCatalogError> {
-        self.item.desc(name)
-    }
-
-    /// Returns the [`mz_sql::func::Func`] associated with this `CatalogEntry`.
-    pub fn func(&self) -> Result<&'static mz_sql::func::Func, SqlCatalogError> {
-        self.item.func(self)
-    }
-
-    /// Returns the inner [`Index`] if this entry is an index, else `None`.
-    pub fn index(&self) -> Option<&Index> {
-        match self.item() {
-            CatalogItem::Index(idx) => Some(idx),
-            _ => None,
-        }
-    }
-
-    /// Returns the inner [`Source`] if this entry is a source, else `None`.
-    pub fn source(&self) -> Option<&Source> {
-        match self.item() {
-            CatalogItem::Source(src) => Some(src),
-            _ => None,
-        }
-    }
-
-    /// Returns the inner [`Sink`] if this entry is a sink, else `None`.
-    pub fn sink(&self) -> Option<&Sink> {
-        match self.item() {
-            CatalogItem::Sink(sink) => Some(sink),
-            _ => None,
-        }
-    }
-
-    /// Returns the inner [`Secret`] if this entry is a secret, else `None`.
-    pub fn secret(&self) -> Option<&Secret> {
-        match self.item() {
-            CatalogItem::Secret(secret) => Some(secret),
-            _ => None,
-        }
-    }
-
-    pub fn connection(&self) -> Result<&Connection, SqlCatalogError> {
-        match self.item() {
-            CatalogItem::Connection(connection) => Ok(connection),
-            _ => {
-                let db_name = match self.name().qualifiers.database_spec {
-                    ResolvedDatabaseSpecifier::Ambient => "".to_string(),
-                    ResolvedDatabaseSpecifier::Id(id) => format!("{id}."),
-                };
-                Err(SqlCatalogError::UnknownConnection(format!(
-                    "{}{}.{}",
-                    db_name,
-                    self.name().qualifiers.schema_spec,
-                    self.name().item
-                )))
-            }
-        }
-    }
-
-    /// Returns the [`mz_storage_types::sources::SourceDesc`] associated with
-    /// this `CatalogEntry`, if any.
-    pub fn source_desc(
-        &self,
-    ) -> Result<Option<&SourceDesc<ReferencedConnection>>, SqlCatalogError> {
-        self.item.source_desc(self)
-    }
-
-    /// Reports whether this catalog entry is a connection.
-    pub fn is_connection(&self) -> bool {
-        matches!(self.item(), CatalogItem::Connection(_))
-    }
-
-    /// Reports whether this catalog entry is a table.
-    pub fn is_table(&self) -> bool {
-        matches!(self.item(), CatalogItem::Table(_))
-    }
-
-    /// Reports whether this catalog entry is a source. Note that this includes
-    /// subsources.
-    pub fn is_source(&self) -> bool {
-        matches!(self.item(), CatalogItem::Source(_))
-    }
-
-    /// Reports whether this catalog entry is a subsource.
-    pub fn is_subsource(&self) -> bool {
-        match &self.item() {
-            CatalogItem::Source(source) => matches!(
-                &source.data_source,
-                DataSourceDesc::Progress | DataSourceDesc::Source
-            ),
-            _ => false,
-        }
-    }
-
-    /// Returns the `GlobalId` of all of this entry's subsources.
-    pub fn subsources(&self) -> BTreeSet<GlobalId> {
-        match &self.item() {
-            CatalogItem::Source(source) => match &source.data_source {
-                DataSourceDesc::Ingestion(ingestion) => ingestion
-                    .source_exports
-                    .keys()
-                    .filter(|id| id != &&self.id)
-                    .copied()
-                    .chain(std::iter::once(ingestion.remap_collection_id))
-                    .collect(),
-                DataSourceDesc::Introspection(_)
-                | DataSourceDesc::Webhook { .. }
-                | DataSourceDesc::Progress
-                | DataSourceDesc::Source => BTreeSet::new(),
-            },
-            CatalogItem::Table(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Index(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => BTreeSet::new(),
-        }
-    }
-
-    /// Returns the `GlobalId` of all of this entry's progress ID.
-    pub fn progress_id(&self) -> Option<GlobalId> {
-        match &self.item() {
-            CatalogItem::Source(source) => match &source.data_source {
-                DataSourceDesc::Ingestion(ingestion) => Some(ingestion.remap_collection_id),
-                DataSourceDesc::Introspection(_)
-                | DataSourceDesc::Progress
-                | DataSourceDesc::Webhook { .. }
-                | DataSourceDesc::Source => None,
-            },
-            CatalogItem::Table(_)
-            | CatalogItem::Log(_)
-            | CatalogItem::View(_)
-            | CatalogItem::MaterializedView(_)
-            | CatalogItem::Sink(_)
-            | CatalogItem::Index(_)
-            | CatalogItem::Type(_)
-            | CatalogItem::Func(_)
-            | CatalogItem::Secret(_)
-            | CatalogItem::Connection(_) => None,
-        }
-    }
-
-    /// Reports whether this catalog entry is a sink.
-    pub fn is_sink(&self) -> bool {
-        matches!(self.item(), CatalogItem::Sink(_))
-    }
-
-    /// Reports whether this catalog entry is a materialized view.
-    pub fn is_materialized_view(&self) -> bool {
-        matches!(self.item(), CatalogItem::MaterializedView(_))
-    }
-
-    /// Reports whether this catalog entry is a secret.
-    pub fn is_secret(&self) -> bool {
-        matches!(self.item(), CatalogItem::Secret(_))
-    }
-
-    /// Reports whether this catalog entry is an introspection source.
-    pub fn is_introspection_source(&self) -> bool {
-        matches!(self.item(), CatalogItem::Log(_))
-    }
-
-    /// Reports whether this catalog entry is an index.
-    pub fn is_index(&self) -> bool {
-        matches!(self.item(), CatalogItem::Index(_))
-    }
-
-    /// Reports whether this catalog entry can be treated as a relation, it can produce rows.
-    pub fn is_relation(&self) -> bool {
-        mz_sql::catalog::ObjectType::from(self.item_type()).is_relation()
-    }
-
-    /// Collects the identifiers of the objects that were encountered when
-    /// resolving names in the item's DDL statement.
-    pub fn uses(&self) -> &ResolvedIds {
-        self.item.uses()
-    }
-
-    /// Returns the `CatalogItem` associated with this catalog entry.
-    pub fn item(&self) -> &CatalogItem {
-        &self.item
-    }
-
-    /// Returns the global ID of this catalog entry.
-    pub fn id(&self) -> GlobalId {
-        self.id
-    }
-
-    /// Returns the OID of this catalog entry.
-    pub fn oid(&self) -> u32 {
-        self.oid
-    }
-
-    /// Returns the fully qualified name of this catalog entry.
-    pub fn name(&self) -> &QualifiedItemName {
-        &self.name
-    }
-
-    /// Returns the identifiers of the dataflows that depend upon this dataflow.
-    pub fn used_by(&self) -> &[GlobalId] {
-        &self.used_by
-    }
-
-    /// Returns the connection ID that this item belongs to, if this item is
-    /// temporary.
-    pub fn conn_id(&self) -> Option<&ConnectionId> {
-        self.item.conn_id()
-    }
-
-    /// Returns the role ID of the entry owner.
-    pub fn owner_id(&self) -> &RoleId {
-        &self.owner_id
-    }
-
-    /// Returns the privileges of the entry.
-    pub fn privileges(&self) -> &PrivilegeMap {
-        &self.privileges
-    }
-}
-
-struct AllocatedBuiltinSystemIds<T> {
-    all_builtins: Vec<(T, GlobalId)>,
-    new_builtins: Vec<(T, GlobalId)>,
-    migrated_builtins: Vec<GlobalId>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct CommentsMap {
-    map: BTreeMap<CommentObjectId, BTreeMap<Option<usize>, String>>,
-}
-
-impl CommentsMap {
-    pub fn update_comment(
-        &mut self,
-        object_id: CommentObjectId,
-        sub_component: Option<usize>,
-        comment: Option<String>,
-    ) -> Option<String> {
-        let object_comments = self.map.entry(object_id).or_default();
-
-        // Either replace the existing comment, or remove it if comment is None/NULL.
-        let (empty, prev) = if let Some(comment) = comment {
-            let prev = object_comments.insert(sub_component, comment);
-            (false, prev)
-        } else {
-            let prev = object_comments.remove(&sub_component);
-            (object_comments.is_empty(), prev)
-        };
-
-        // Cleanup entries that are now empty.
-        if empty {
-            self.map.remove(&object_id);
-        }
-
-        // Return the previous comment, if there was one, for easy removal.
-        prev
-    }
-
-    /// Remove all comments for `object_id` from the map.
-    ///
-    /// Generally there is one comment for a given [`CommentObjectId`], but in the case of
-    /// relations you can also have comments on the individual columns. Dropping the comments for a
-    /// relation will also drop all of the comments on any columns.
-    pub fn drop_comments(
-        &mut self,
-        object_id: CommentObjectId,
-    ) -> Vec<(CommentObjectId, Option<usize>, String)> {
-        match self.map.remove(&object_id) {
-            None => Vec::new(),
-            Some(comments) => comments
-                .into_iter()
-                .map(|(sub_comp, comment)| (object_id, sub_comp, comment))
-                .collect(),
-        }
-    }
-
-    pub fn iter(&self) -> impl Iterator<Item = (CommentObjectId, Option<usize>, &str)> {
-        self.map
-            .iter()
-            .map(|(id, comments)| {
-                comments
-                    .iter()
-                    .map(|(pos, comment)| (*id, *pos, comment.as_str()))
-            })
-            .flatten()
-    }
-}
-
-impl Serialize for CommentsMap {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        let comment_count = self
-            .map
-            .iter()
-            .map(|(_object_id, comments)| comments.len())
-            .sum();
-
-        let mut seq = serializer.serialize_seq(Some(comment_count))?;
-        for (object_id, sub) in &self.map {
-            for (sub_component, comment) in sub {
-                seq.serialize_element(&(
-                    format!("{object_id:?}"),
-                    format!("{sub_component:?}"),
-                    comment,
-                ))?;
-            }
-        }
-        seq.end()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Default)]
-pub struct DefaultPrivileges {
-    #[serde(serialize_with = "mz_ore::serde::map_key_to_string")]
-    privileges: BTreeMap<DefaultPrivilegeObject, RoleDefaultPrivileges>,
-}
-
-// Use a new type here because otherwise we have two levels of BTreeMap, both needing
-// map_key_to_string.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Default)]
-struct RoleDefaultPrivileges(
-    /// Denormalized, the key is the grantee Role.
-    #[serde(serialize_with = "mz_ore::serde::map_key_to_string")]
-    BTreeMap<RoleId, DefaultPrivilegeAclItem>,
-);
-
-impl Deref for RoleDefaultPrivileges {
-    type Target = BTreeMap<RoleId, DefaultPrivilegeAclItem>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for RoleDefaultPrivileges {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl DefaultPrivileges {
-    /// Add a new default privilege into the set of all default privileges.
-    fn grant(&mut self, object: DefaultPrivilegeObject, privilege: DefaultPrivilegeAclItem) {
-        if privilege.acl_mode.is_empty() {
-            return;
-        }
-
-        let privileges = self.privileges.entry(object).or_default();
-        if let Some(default_privilege) = privileges.get_mut(&privilege.grantee) {
-            default_privilege.acl_mode |= privilege.acl_mode;
-        } else {
-            privileges.insert(privilege.grantee, privilege);
-        }
-    }
-
-    /// Revoke a default privilege from the set of all default privileges.
-    fn revoke(&mut self, object: &DefaultPrivilegeObject, privilege: &DefaultPrivilegeAclItem) {
-        if let Some(privileges) = self.privileges.get_mut(object) {
-            if let Some(default_privilege) = privileges.get_mut(&privilege.grantee) {
-                default_privilege.acl_mode =
-                    default_privilege.acl_mode.difference(privilege.acl_mode);
-                if default_privilege.acl_mode.is_empty() {
-                    privileges.remove(&privilege.grantee);
-                }
-            }
-            if privileges.is_empty() {
-                self.privileges.remove(object);
-            }
-        }
-    }
-
-    /// Get the privileges that will be granted on all objects matching `object` to `grantee`, if
-    /// any exist.
-    fn get_privileges_for_grantee(
-        &self,
-        object: &DefaultPrivilegeObject,
-        grantee: &RoleId,
-    ) -> Option<&AclMode> {
-        self.privileges
-            .get(object)
-            .and_then(|privileges| privileges.get(grantee))
-            .map(|privilege| &privilege.acl_mode)
-    }
-
-    /// Get all default privileges that apply to the provided object details.
-    fn get_applicable_privileges(
-        &self,
-        role_id: RoleId,
-        database_id: Option<DatabaseId>,
-        schema_id: Option<SchemaId>,
-        object_type: mz_sql::catalog::ObjectType,
-    ) -> impl Iterator<Item = DefaultPrivilegeAclItem> + '_ {
-        // Privileges consider all relations to be of type table due to PostgreSQL compatibility. We
-        // don't require the caller to worry about that and we will map their `object_type` to the
-        // correct type for privileges.
-        let privilege_object_type = if object_type.is_relation() {
-            mz_sql::catalog::ObjectType::Table
-        } else {
-            object_type
-        };
-        let valid_acl_mode = rbac::all_object_privileges(SystemObjectType::Object(object_type));
-
-        // Collect all entries that apply to the provided object details.
-        // If either `database_id` or `schema_id` are `None`, then we might end up with duplicate
-        // entries in the vec below. That's OK because we consolidate the results after.
-        [
-            DefaultPrivilegeObject {
-                role_id,
-                database_id,
-                schema_id,
-                object_type: privilege_object_type,
-            },
-            DefaultPrivilegeObject {
-                role_id,
-                database_id,
-                schema_id: None,
-                object_type: privilege_object_type,
-            },
-            DefaultPrivilegeObject {
-                role_id,
-                database_id: None,
-                schema_id: None,
-                object_type: privilege_object_type,
-            },
-            DefaultPrivilegeObject {
-                role_id: RoleId::Public,
-                database_id,
-                schema_id,
-                object_type: privilege_object_type,
-            },
-            DefaultPrivilegeObject {
-                role_id: RoleId::Public,
-                database_id,
-                schema_id: None,
-                object_type: privilege_object_type,
-            },
-            DefaultPrivilegeObject {
-                role_id: RoleId::Public,
-                database_id: None,
-                schema_id: None,
-                object_type: privilege_object_type,
-            },
-        ]
-        .into_iter()
-        .filter_map(|object| self.privileges.get(&object))
-        .flat_map(|acl_map| acl_map.values())
-        // Consolidate privileges with a common grantee.
-        .fold(
-            BTreeMap::new(),
-            |mut accum, DefaultPrivilegeAclItem { grantee, acl_mode }| {
-                let accum_acl_mode = accum.entry(grantee).or_insert_with(AclMode::empty);
-                *accum_acl_mode |= *acl_mode;
-                accum
-            },
-        )
-        .into_iter()
-        // Restrict the acl_mode to only privileges valid for the provided object type. If the
-        // default privilege has an object type of Table, then it may contain privileges valid for
-        // tables but not other relations. If the passed in object type is another relation, then
-        // we need to remove any privilege that is not valid for the specified relation.
-        .map(move |(grantee, acl_mode)| (grantee, acl_mode & valid_acl_mode))
-        // Filter out empty privileges.
-        .filter(|(_, acl_mode)| !acl_mode.is_empty())
-        .map(|(grantee, acl_mode)| DefaultPrivilegeAclItem {
-            grantee: *grantee,
-            acl_mode,
-        })
-    }
-
-    fn iter(
-        &self,
-    ) -> impl Iterator<
-        Item = (
-            &DefaultPrivilegeObject,
-            impl Iterator<Item = &DefaultPrivilegeAclItem>,
-        ),
-    > {
-        self.privileges
-            .iter()
-            .map(|(object, acl_map)| (object, acl_map.values()))
-    }
-}
-
-#[derive(Debug)]
-pub enum CatalogItemRebuilder {
-    SystemSource(CatalogItem),
-    Object {
-        id: GlobalId,
-        sql: String,
-        is_retained_metrics_object: bool,
-        custom_logical_compaction_window: Option<Duration>,
-    },
-}
-
-impl CatalogItemRebuilder {
-    fn new(
-        entry: &CatalogEntry,
-        id: GlobalId,
-        ancestor_ids: &BTreeMap<GlobalId, GlobalId>,
-    ) -> Self {
-        if id.is_system()
-            && (entry.is_table() || entry.is_introspection_source() || entry.is_source())
-        {
-            Self::SystemSource(entry.item().clone())
-        } else {
-            let create_sql = entry.create_sql().to_string();
-            assert_ne!(create_sql.to_lowercase(), CREATE_SQL_TODO.to_lowercase());
-            let mut create_stmt = mz_sql::parse::parse(&create_sql)
-                .expect("invalid create sql persisted to catalog")
-                .into_element()
-                .ast;
-            mz_sql::ast::transform::create_stmt_replace_ids(&mut create_stmt, ancestor_ids);
-            Self::Object {
-                id,
-                sql: create_stmt.to_ast_string_stable(),
-                is_retained_metrics_object: entry.item().is_retained_metrics_object(),
-                custom_logical_compaction_window: entry.item().custom_logical_compaction_window(),
-            }
-        }
-    }
-
-    fn build(self, catalog: &Catalog) -> CatalogItem {
-        match self {
-            Self::SystemSource(item) => item,
-            Self::Object {
-                id,
-                sql,
-                is_retained_metrics_object,
-                custom_logical_compaction_window,
-            } => catalog
-                .parse_item(
-                    id,
-                    sql.clone(),
-                    None,
-                    is_retained_metrics_object,
-                    custom_logical_compaction_window,
-                )
-                .unwrap_or_else(|error| panic!("invalid persisted create sql ({error:?}): {sql}")),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct BuiltinMigrationMetadata {
-    // Used to drop objects on STORAGE nodes
-    pub previous_sink_ids: Vec<GlobalId>,
-    pub previous_materialized_view_ids: Vec<GlobalId>,
-    pub previous_source_ids: Vec<GlobalId>,
-    // Used to update in memory catalog state
-    pub all_drop_ops: Vec<GlobalId>,
-    pub all_create_ops: Vec<(
-        GlobalId,
-        u32,
-        QualifiedItemName,
-        RoleId,
-        PrivilegeMap,
-        CatalogItemRebuilder,
-    )>,
-    pub introspection_source_index_updates:
-        BTreeMap<ClusterId, Vec<(LogVariant, String, GlobalId)>>,
-    // Used to update persisted on disk catalog state
-    pub migrated_system_object_mappings: BTreeMap<GlobalId, SystemObjectMapping>,
-    pub user_drop_ops: Vec<GlobalId>,
-    pub user_create_ops: Vec<(GlobalId, SchemaId, String)>,
-}
-
-impl BuiltinMigrationMetadata {
-    fn new() -> BuiltinMigrationMetadata {
-        BuiltinMigrationMetadata {
-            previous_sink_ids: Vec::new(),
-            previous_materialized_view_ids: Vec::new(),
-            previous_source_ids: Vec::new(),
-            all_drop_ops: Vec::new(),
-            all_create_ops: Vec::new(),
-            introspection_source_index_updates: BTreeMap::new(),
-            migrated_system_object_mappings: BTreeMap::new(),
-            user_drop_ops: Vec::new(),
-            user_create_ops: Vec::new(),
-        }
-    }
-}
-
 impl Catalog {
-    /// Opens or creates a catalog that stores data at `path`.
-    ///
-    /// Returns the catalog, metadata about builtin objects that have changed
-    /// schemas since last restart, a list of updates to builtin tables that
-    /// describe the initial state of the catalog, and the version of the
-    /// catalog before any migrations were performed.
-    #[tracing::instrument(name = "catalog::open", level = "info", skip_all)]
-    pub async fn open(
-        config: Config<'_>,
-    ) -> Result<
-        (
-            Catalog,
-            BuiltinMigrationMetadata,
-            Vec<BuiltinTableUpdate>,
-            String,
-        ),
-        AdapterError,
-    > {
-        for builtin_role in BUILTIN_ROLES {
-            assert!(
-                is_reserved_name(builtin_role.name),
-                "builtin role {builtin_role:?} must start with one of the following prefixes {}",
-                BUILTIN_PREFIXES.join(", ")
-            );
-        }
-        for builtin_cluster in BUILTIN_CLUSTERS {
-            assert!(
-                is_reserved_name(builtin_cluster.name),
-                "builtin cluster {builtin_cluster:?} must start with one of the following prefixes {}",
-                BUILTIN_PREFIXES.join(", ")
-            );
-        }
-
-        let mut catalog = Catalog {
-            state: CatalogState {
-                database_by_name: BTreeMap::new(),
-                database_by_id: BTreeMap::new(),
-                entry_by_id: BTreeMap::new(),
-                ambient_schemas_by_name: BTreeMap::new(),
-                ambient_schemas_by_id: BTreeMap::new(),
-                temporary_schemas: BTreeMap::new(),
-                clusters_by_id: BTreeMap::new(),
-                clusters_by_name: BTreeMap::new(),
-                clusters_by_linked_object_id: BTreeMap::new(),
-                roles_by_name: BTreeMap::new(),
-                roles_by_id: BTreeMap::new(),
-                config: mz_sql::catalog::CatalogConfig {
-                    start_time: to_datetime((config.now)()),
-                    start_instant: Instant::now(),
-                    nonce: rand::random(),
-                    environment_id: config.environment_id,
-                    session_id: Uuid::new_v4(),
-                    build_info: config.build_info,
-                    timestamp_interval: Duration::from_secs(1),
-                    now: config.now.clone(),
-                },
-                oid_counter: FIRST_USER_OID,
-                cluster_replica_sizes: config.cluster_replica_sizes,
-                default_storage_cluster_size: config.default_storage_cluster_size,
-                availability_zones: config.availability_zones,
-                system_configuration: {
-                    let mut s = SystemVars::new(config.active_connection_count)
-                        .set_unsafe(config.unsafe_mode);
-                    if config.all_features {
-                        s.enable_all_feature_flags_by_default();
-                    }
-                    s
-                },
-                egress_ips: config.egress_ips,
-                aws_principal_context: config.aws_principal_context,
-                aws_privatelink_availability_zones: config.aws_privatelink_availability_zones,
-                http_host_name: config.http_host_name,
-                default_privileges: DefaultPrivileges::default(),
-                system_privileges: PrivilegeMap::default(),
-                comments: CommentsMap::default(),
-            },
-            plans: CatalogPlans {
-                optimized_plan_by_id: Default::default(),
-                physical_plan_by_id: Default::default(),
-                dataflow_metainfos: BTreeMap::new(),
-            },
-            transient_revision: 0,
-            storage: Arc::new(tokio::sync::Mutex::new(config.storage)),
-        };
-
-        // Choose a time at which to boot. This is the time at which we will run
-        // internal migrations.
-        //
-        // This time is usually the current system time, but with protection
-        // against backwards time jumps, even across restarts.
-        let boot_ts = {
-            let mut storage = catalog.storage().await;
-            let previous_ts = storage
-                .get_timestamp(&Timeline::EpochMilliseconds)
-                .await?
-                .expect("missing EpochMilliseconds timeline");
-            let boot_ts = timestamp_oracle::catalog_oracle::monotonic_now(config.now, previous_ts);
-            if !storage.is_read_only() {
-                // IMPORTANT: we durably record the new timestamp before using it.
-                storage
-                    .set_timestamp(&Timeline::EpochMilliseconds, boot_ts)
-                    .await?;
-            }
-
-            boot_ts
-        };
-
-        catalog.create_temporary_schema(&SYSTEM_CONN_ID, MZ_SYSTEM_ROLE_ID)?;
-
-        let databases = catalog.storage().await.get_databases().await?;
-        for mz_catalog::Database {
-            id,
-            name,
-            owner_id,
-            privileges,
-        } in databases
-        {
-            let oid = catalog.allocate_oid()?;
-            catalog.state.database_by_id.insert(
-                id.clone(),
-                Database {
-                    name: name.clone(),
-                    id,
-                    oid,
-                    schemas_by_id: BTreeMap::new(),
-                    schemas_by_name: BTreeMap::new(),
-                    owner_id,
-                    privileges: PrivilegeMap::from_mz_acl_items(privileges),
-                },
-            );
-            catalog
-                .state
-                .database_by_name
-                .insert(name.clone(), id.clone());
-        }
-
-        let schemas = catalog.storage().await.get_schemas().await?;
-        for mz_catalog::Schema {
-            id,
-            name,
-            database_id,
-            owner_id,
-            privileges,
-        } in schemas
-        {
-            let oid = catalog.allocate_oid()?;
-            let (schemas_by_id, schemas_by_name, database_spec) = match &database_id {
-                Some(database_id) => {
-                    let db = catalog
-                        .state
-                        .database_by_id
-                        .get_mut(database_id)
-                        .expect("catalog out of sync");
-                    (
-                        &mut db.schemas_by_id,
-                        &mut db.schemas_by_name,
-                        ResolvedDatabaseSpecifier::Id(*database_id),
-                    )
-                }
-                None => (
-                    &mut catalog.state.ambient_schemas_by_id,
-                    &mut catalog.state.ambient_schemas_by_name,
-                    ResolvedDatabaseSpecifier::Ambient,
-                ),
-            };
-            schemas_by_id.insert(
-                id.clone(),
-                Schema {
-                    name: QualifiedSchemaName {
-                        database: database_spec,
-                        schema: name.clone(),
-                    },
-                    id: SchemaSpecifier::Id(id.clone()),
-                    oid,
-                    items: BTreeMap::new(),
-                    functions: BTreeMap::new(),
-                    owner_id,
-                    privileges: PrivilegeMap::from_mz_acl_items(privileges),
-                },
-            );
-            schemas_by_name.insert(name.clone(), id);
-        }
-
-        let roles = catalog.storage().await.get_roles().await?;
-        for mz_catalog::Role {
-            id,
-            name,
-            attributes,
-            membership,
-            vars,
-        } in roles
-        {
-            let oid = catalog.allocate_oid()?;
-            catalog.state.roles_by_name.insert(name.clone(), id);
-            catalog.state.roles_by_id.insert(
-                id,
-                Role {
-                    name,
-                    id,
-                    oid,
-                    attributes,
-                    membership,
-                    vars,
-                },
-            );
-        }
-
-        let default_privileges = catalog.storage().await.get_default_privileges().await?;
-        for mz_catalog::DefaultPrivilege { object, acl_item } in default_privileges {
-            catalog.state.default_privileges.grant(object, acl_item);
-        }
-
-        let system_privileges = catalog.storage().await.get_system_privileges().await?;
-        catalog.state.system_privileges.grant_all(system_privileges);
-
-        catalog
-            .load_system_configuration(
-                config.system_parameter_defaults,
-                config.system_parameter_sync_config,
-                boot_ts,
-            )
-            .await?;
-        // We need to set this variable ASAP, so that builtins get planned with the correct value
-        let variable_length_row_encoding = catalog
-            .system_config()
-            .variable_length_row_encoding_DANGEROUS();
-        mz_repr::VARIABLE_LENGTH_ROW_ENCODING
-            .store(variable_length_row_encoding, atomic::Ordering::SeqCst);
-
-        let comments = catalog.storage().await.get_comments().await?;
-        for mz_catalog::Comment {
-            object_id,
-            sub_component,
-            comment,
-        } in comments
-        {
-            catalog
-                .state
-                .comments
-                .update_comment(object_id, sub_component, Some(comment));
-        }
-
-        // Now that LD is loaded, set the intended stash timeout.
-        // TODO: Move this into the stash constructor.
-        catalog
-            .storage()
-            .await
-            .set_connect_timeout(catalog.system_config().crdb_connect_timeout())
-            .await;
-
-        catalog.load_builtin_types().await?;
-
-        let persisted_builtin_ids: BTreeMap<_, _> = catalog
-            .storage()
-            .await
-            .get_system_items()
-            .await?
-            .into_iter()
-            .map(|mapping| (mapping.description, mapping.unique_identifier))
-            .collect();
-        let AllocatedBuiltinSystemIds {
-            all_builtins,
-            new_builtins,
-            migrated_builtins,
-        } = catalog
-            .allocate_system_ids(
-                BUILTINS::iter()
-                    .filter(|builtin| !matches!(builtin, Builtin::Type(_)))
-                    .collect(),
-                |builtin| {
-                    persisted_builtin_ids
-                        .get(&SystemObjectDescription {
-                            schema_name: builtin.schema().to_string(),
-                            object_type: builtin.catalog_item_type(),
-                            object_name: builtin.name().to_string(),
-                        })
-                        .cloned()
-                },
-            )
-            .await?;
-
-        let id_fingerprint_map: BTreeMap<GlobalId, String> = all_builtins
-            .iter()
-            .map(|(builtin, id)| (*id, builtin.fingerprint()))
-            .collect();
-        let (builtin_indexes, builtin_non_indexes): (Vec<_>, Vec<_>) = all_builtins
-            .into_iter()
-            .partition(|(builtin, _)| matches!(builtin, Builtin::Index(_)));
-
-        {
-            let span = tracing::span!(tracing::Level::DEBUG, "builtin_non_indexes");
-            let _enter = span.enter();
-            for (builtin, id) in builtin_non_indexes {
-                let schema_id = catalog.state.ambient_schemas_by_name[builtin.schema()];
-                let name = QualifiedItemName {
-                    qualifiers: ItemQualifiers {
-                        database_spec: ResolvedDatabaseSpecifier::Ambient,
-                        schema_spec: SchemaSpecifier::Id(schema_id),
-                    },
-                    item: builtin.name().into(),
-                };
-                match builtin {
-                    Builtin::Log(log) => {
-                        let oid = catalog.allocate_oid()?;
-                        catalog.state.insert_item(
-                            id,
-                            oid,
-                            name.clone(),
-                            CatalogItem::Log(Log {
-                                variant: log.variant.clone(),
-                                has_storage_collection: false,
-                            }),
-                            MZ_SYSTEM_ROLE_ID,
-                            PrivilegeMap::from_mz_acl_items(vec![
-                                rbac::default_builtin_object_privilege(
-                                    mz_sql::catalog::ObjectType::Source,
-                                ),
-                                rbac::owner_privilege(
-                                    mz_sql::catalog::ObjectType::Source,
-                                    MZ_SYSTEM_ROLE_ID,
-                                ),
-                            ]),
-                        );
-                    }
-
-                    Builtin::Table(table) => {
-                        let oid = catalog.allocate_oid()?;
-                        catalog.state.insert_item(
-                            id,
-                            oid,
-                            name.clone(),
-                            CatalogItem::Table(Table {
-                                create_sql: CREATE_SQL_TODO.to_string(),
-                                desc: table.desc.clone(),
-                                defaults: vec![Expr::null(); table.desc.arity()],
-                                conn_id: None,
-                                resolved_ids: ResolvedIds(BTreeSet::new()),
-                                custom_logical_compaction_window: table
-                                    .is_retained_metrics_object
-                                    .then(|| catalog.state.system_config().metrics_retention()),
-                                is_retained_metrics_object: table.is_retained_metrics_object,
-                            }),
-                            MZ_SYSTEM_ROLE_ID,
-                            PrivilegeMap::from_mz_acl_items(vec![
-                                rbac::default_builtin_object_privilege(
-                                    mz_sql::catalog::ObjectType::Table,
-                                ),
-                                rbac::owner_privilege(
-                                    mz_sql::catalog::ObjectType::Table,
-                                    MZ_SYSTEM_ROLE_ID,
-                                ),
-                            ]),
-                        );
-                    }
-                    Builtin::Index(_) => {
-                        unreachable!("handled later once clusters have been created")
-                    }
-                    Builtin::View(view) => {
-                        let item = catalog
-                        .parse_item(
-                            id,
-                            view.sql.into(),
-                            None,
-                            false,
-                            None
-                        )
-                        .unwrap_or_else(|e| {
-                            panic!(
-                                "internal error: failed to load bootstrap view:\n\
-                                    {}\n\
-                                    error:\n\
-                                    {:?}\n\n\
-                                    make sure that the schema name is specified in the builtin view's create sql statement.",
-                                view.name, e
-                            )
-                        });
-                        let oid = catalog.allocate_oid()?;
-                        catalog.state.insert_item(
-                            id,
-                            oid,
-                            name,
-                            item,
-                            MZ_SYSTEM_ROLE_ID,
-                            PrivilegeMap::from_mz_acl_items(vec![
-                                rbac::default_builtin_object_privilege(
-                                    mz_sql::catalog::ObjectType::View,
-                                ),
-                                rbac::owner_privilege(
-                                    mz_sql::catalog::ObjectType::View,
-                                    MZ_SYSTEM_ROLE_ID,
-                                ),
-                            ]),
-                        );
-                    }
-
-                    Builtin::Type(_) => unreachable!("loaded separately"),
-
-                    Builtin::Func(func) => {
-                        let oid = catalog.allocate_oid()?;
-                        catalog.state.insert_item(
-                            id,
-                            oid,
-                            name.clone(),
-                            CatalogItem::Func(Func { inner: func.inner }),
-                            MZ_SYSTEM_ROLE_ID,
-                            PrivilegeMap::default(),
-                        );
-                    }
-
-                    Builtin::Source(coll) => {
-                        let introspection_type = match &coll.data_source {
-                            Some(i) => i.clone(),
-                            None => continue,
-                        };
-
-                        let oid = catalog.allocate_oid()?;
-                        catalog.state.insert_item(
-                            id,
-                            oid,
-                            name.clone(),
-                            CatalogItem::Source(Source {
-                                create_sql: CREATE_SQL_TODO.to_string(),
-                                data_source: DataSourceDesc::Introspection(introspection_type),
-                                desc: coll.desc.clone(),
-                                timeline: Timeline::EpochMilliseconds,
-                                resolved_ids: ResolvedIds(BTreeSet::new()),
-                                custom_logical_compaction_window: coll
-                                    .is_retained_metrics_object
-                                    .then(|| catalog.state.system_config().metrics_retention()),
-                                is_retained_metrics_object: coll.is_retained_metrics_object,
-                            }),
-                            MZ_SYSTEM_ROLE_ID,
-                            PrivilegeMap::from_mz_acl_items(vec![
-                                rbac::default_builtin_object_privilege(
-                                    mz_sql::catalog::ObjectType::Source,
-                                ),
-                                rbac::owner_privilege(
-                                    mz_sql::catalog::ObjectType::Source,
-                                    MZ_SYSTEM_ROLE_ID,
-                                ),
-                            ]),
-                        );
-                    }
-                }
-            }
-        }
-
-        let clusters = catalog.storage().await.get_clusters().await?;
-        let mut cluster_azs = BTreeMap::new();
-        for mz_catalog::Cluster {
-            id,
-            name,
-            linked_object_id,
-            owner_id,
-            privileges,
-            config,
-        } in clusters
-        {
-            let introspection_source_index_gids = catalog
-                .storage()
-                .await
-                .get_introspection_source_indexes(id)
-                .await?;
-
-            let AllocatedBuiltinSystemIds {
-                all_builtins: all_indexes,
-                new_builtins: new_indexes,
-                ..
-            } = catalog
-                .allocate_system_ids(BUILTINS::logs().collect(), |log| {
-                    introspection_source_index_gids
-                        .get(log.name)
-                        .cloned()
-                        // We migrate introspection sources later so we can hardcode the fingerprint as ""
-                        .map(|id| SystemObjectUniqueIdentifier {
-                            id,
-                            fingerprint: "".to_string(),
-                        })
-                })
-                .await?;
-
-            catalog
-                .storage()
-                .await
-                .set_introspection_source_indexes(
-                    new_indexes
-                        .iter()
-                        .map(|(log, index_id)| (id, log.name, *index_id))
-                        .collect(),
-                )
-                .await?;
-
-            if let mz_catalog::ClusterVariant::Managed(managed) = &config.variant {
-                cluster_azs.insert(id, managed.availability_zones.clone());
-            }
-
-            catalog.state.insert_cluster(
-                id,
-                name,
-                linked_object_id,
-                all_indexes,
-                owner_id,
-                PrivilegeMap::from_mz_acl_items(privileges),
-                config.into(),
-            );
-        }
-
-        let replicas = catalog.storage().await.get_cluster_replicas().await?;
-        for mz_catalog::ClusterReplica {
-            cluster_id,
-            replica_id,
-            name,
-            config,
-            owner_id,
-        } in replicas
-        {
-            let logging = ReplicaLogging {
-                log_logging: config.logging.log_logging,
-                interval: config.logging.interval,
-            };
-            let config = ReplicaConfig {
-                location: catalog.concretize_replica_location(
-                    config.location,
-                    &vec![],
-                    cluster_azs.get(&cluster_id).map(|zones| &**zones),
-                )?,
-                compute: ComputeReplicaConfig {
-                    logging,
-                    idle_arrangement_merge_effort: config.idle_arrangement_merge_effort,
-                },
-            };
-
-            // And write the allocated sources back to storage
-            catalog
-                .storage()
-                .await
-                .set_replica_config(
-                    replica_id,
-                    cluster_id,
-                    name.clone(),
-                    config.clone().into(),
-                    owner_id,
-                )
-                .await?;
-
-            catalog
-                .state
-                .insert_cluster_replica(cluster_id, name, replica_id, config, owner_id);
-        }
-
-        for (builtin, id) in builtin_indexes {
-            let schema_id = catalog.state.ambient_schemas_by_name[builtin.schema()];
-            let name = QualifiedItemName {
-                qualifiers: ItemQualifiers {
-                    database_spec: ResolvedDatabaseSpecifier::Ambient,
-                    schema_spec: SchemaSpecifier::Id(schema_id),
-                },
-                item: builtin.name().into(),
-            };
-            match builtin {
-                Builtin::Index(index) => {
-                    let mut item = catalog
-                        .parse_item(
-                            id,
-                            index.sql.into(),
-                            None,
-                            index.is_retained_metrics_object,
-                            if index.is_retained_metrics_object { Some(catalog.state.system_config().metrics_retention())} else { None },
-                        )
-                        .unwrap_or_else(|e| {
-                            panic!(
-                                "internal error: failed to load bootstrap index:\n\
-                                    {}\n\
-                                    error:\n\
-                                    {:?}\n\n\
-                                    make sure that the schema name is specified in the builtin index's create sql statement.",
-                                index.name, e
-                            )
-                        });
-                    let CatalogItem::Index(_) = &mut item else {
-                        panic!("internal error: builtin index {}'s SQL does not begin with \"CREATE INDEX\".", index.name);
-                    };
-
-                    let oid = catalog.allocate_oid()?;
-                    catalog.state.insert_item(
-                        id,
-                        oid,
-                        name,
-                        item,
-                        MZ_SYSTEM_ROLE_ID,
-                        PrivilegeMap::default(),
-                    );
-                }
-                Builtin::Log(_)
-                | Builtin::Table(_)
-                | Builtin::View(_)
-                | Builtin::Type(_)
-                | Builtin::Func(_)
-                | Builtin::Source(_) => {
-                    unreachable!("handled above")
-                }
-            }
-        }
-
-        let new_system_id_mappings = new_builtins
-            .iter()
-            .map(|(builtin, id)| SystemObjectMapping {
-                description: SystemObjectDescription {
-                    schema_name: builtin.schema().to_string(),
-                    object_type: builtin.catalog_item_type(),
-                    object_name: builtin.name().to_string(),
-                },
-                unique_identifier: SystemObjectUniqueIdentifier {
-                    id: *id,
-                    fingerprint: builtin.fingerprint(),
-                },
-            })
-            .collect();
-        catalog
-            .storage()
-            .await
-            .set_system_items(new_system_id_mappings)
-            .await?;
-
-        let last_seen_version = catalog
-            .storage()
-            .await
-            .get_catalog_content_version()
-            .await?
-            .unwrap_or_else(|| "new".to_string());
-
-        if !config.skip_migrations {
-            migrate::migrate(&mut catalog, config.connection_context)
-                .await
-                .map_err(|e| {
-                    Error::new(ErrorKind::FailedMigration {
-                        last_seen_version: last_seen_version.clone(),
-                        this_version: catalog.config().build_info.version,
-                        cause: e.to_string(),
-                    })
-                })?;
-            catalog
-                .storage()
-                .await
-                .set_catalog_content_version(catalog.config().build_info.version)
-                .await?;
-        }
-
-        let mut catalog = {
-            let mut storage = catalog.storage().await;
-            let mut tx = storage.transaction().await?;
-            let catalog = Self::load_catalog_items(&mut tx, &catalog)?;
-            tx.commit().await?;
-            catalog
-        };
-
-        let mut builtin_migration_metadata = catalog
-            .generate_builtin_migration_metadata(migrated_builtins, id_fingerprint_map)
-            .await?;
-        catalog.apply_in_memory_builtin_migration(&mut builtin_migration_metadata)?;
-        catalog
-            .apply_persisted_builtin_migration(&mut builtin_migration_metadata)
-            .await?;
-
-        // Load public keys for SSH connections from the secrets store to the catalog
-        for (id, entry) in catalog.state.entry_by_id.iter_mut() {
-            if let CatalogItem::Connection(ref mut connection) = entry.item {
-                if let mz_storage_types::connections::Connection::Ssh(ref mut ssh) =
-                    connection.connection
-                {
-                    let secret = config.secrets_reader.read(*id).await?;
-                    let keyset = SshKeyPairSet::from_bytes(&secret)?;
-                    let public_key_pair = keyset.public_keys();
-                    ssh.public_keys = Some(public_key_pair);
-                }
-            }
-        }
-
-        let mut builtin_table_updates = vec![];
-        for (schema_id, schema) in &catalog.state.ambient_schemas_by_id {
-            let db_spec = ResolvedDatabaseSpecifier::Ambient;
-            builtin_table_updates.push(catalog.state.pack_schema_update(&db_spec, schema_id, 1));
-            for (_item_name, item_id) in &schema.items {
-                builtin_table_updates.extend(catalog.state.pack_item_update(*item_id, 1));
-            }
-            for (_item_name, function_id) in &schema.functions {
-                builtin_table_updates.extend(catalog.state.pack_item_update(*function_id, 1));
-            }
-        }
-        for (_id, db) in &catalog.state.database_by_id {
-            builtin_table_updates.push(catalog.state.pack_database_update(db, 1));
-            let db_spec = ResolvedDatabaseSpecifier::Id(db.id.clone());
-            for (schema_id, schema) in &db.schemas_by_id {
-                builtin_table_updates
-                    .push(catalog.state.pack_schema_update(&db_spec, schema_id, 1));
-                for (_item_name, item_id) in &schema.items {
-                    builtin_table_updates.extend(catalog.state.pack_item_update(*item_id, 1));
-                }
-                for (_item_name, function_id) in &schema.functions {
-                    builtin_table_updates.extend(catalog.state.pack_item_update(*function_id, 1));
-                }
-            }
-        }
-        for (id, sub_component, comment) in catalog.state.comments.iter() {
-            builtin_table_updates.push(catalog.state.pack_comment_update(
-                id,
-                sub_component,
-                comment,
-                1,
-            ));
-        }
-        for (_id, role) in &catalog.state.roles_by_id {
-            if let Some(builtin_update) = catalog.state.pack_role_update(role.id, 1) {
-                builtin_table_updates.push(builtin_update);
-            }
-            for group_id in role.membership.map.keys() {
-                builtin_table_updates.push(
-                    catalog
-                        .state
-                        .pack_role_members_update(*group_id, role.id, 1),
-                )
-            }
-        }
-        for (default_privilege_object, default_privilege_acl_items) in
-            catalog.state.default_privileges.iter()
-        {
-            for default_privilege_acl_item in default_privilege_acl_items {
-                builtin_table_updates.push(catalog.state.pack_default_privileges_update(
-                    default_privilege_object,
-                    &default_privilege_acl_item.grantee,
-                    &default_privilege_acl_item.acl_mode,
-                    1,
-                ));
-            }
-        }
-        for system_privilege in catalog.state.system_privileges.all_values_owned() {
-            builtin_table_updates.push(
-                catalog
-                    .state
-                    .pack_system_privileges_update(system_privilege, 1),
-            );
-        }
-        for (id, cluster) in &catalog.state.clusters_by_id {
-            builtin_table_updates.push(catalog.state.pack_cluster_update(&cluster.name, 1));
-            if let Some(linked_object_id) = cluster.linked_object_id {
-                builtin_table_updates.push(catalog.state.pack_cluster_link_update(
-                    &cluster.name,
-                    linked_object_id,
-                    1,
-                ));
-            }
-            for (replica_name, replica_id) in cluster.replicas().map(|r| (&r.name, r.replica_id)) {
-                builtin_table_updates.extend(catalog.state.pack_cluster_replica_update(
-                    *id,
-                    replica_name,
-                    1,
-                ));
-                let replica = catalog.state.get_cluster_replica(*id, replica_id);
-                for process_id in 0..replica.config.location.num_processes() {
-                    let update = catalog.state.pack_cluster_replica_status_update(
-                        *id,
-                        replica_id,
-                        u64::cast_from(process_id),
-                        1,
-                    );
-                    builtin_table_updates.push(update);
-                }
-            }
-        }
-        // Operators aren't stored in the catalog, but we would like them in
-        // introspection views.
-        for (op, func) in OP_IMPLS.iter() {
-            match func {
-                mz_sql::func::Func::Scalar(impls) => {
-                    for imp in impls {
-                        builtin_table_updates.push(catalog.state.pack_op_update(
-                            op,
-                            imp.details(),
-                            1,
-                        ));
-                    }
-                }
-                _ => unreachable!("all operators must be scalar functions"),
-            }
-        }
-        let audit_logs = catalog.storage().await.get_audit_logs().await?;
-        for event in audit_logs {
-            builtin_table_updates.push(catalog.state.pack_audit_log_update(&event)?);
-        }
-
-        // To avoid reading over storage_usage events multiple times, do both
-        // the table updates and delete calculations in a single read over the
-        // data.
-        let storage_usage_events = catalog
-            .storage()
-            .await
-            .get_and_prune_storage_usage(config.storage_usage_retention_period, boot_ts)
-            .await?;
-        for event in storage_usage_events {
-            builtin_table_updates.push(catalog.state.pack_storage_usage_update(&event)?);
-        }
-
-        for ip in &catalog.state.egress_ips {
-            builtin_table_updates.push(catalog.state.pack_egress_ip_update(ip)?);
-        }
-
-        Ok((
-            catalog,
-            builtin_migration_metadata,
-            builtin_table_updates,
-            last_seen_version,
-        ))
-    }
-
-    /// Loads the system configuration from the various locations in which its
-    /// values and value overrides can reside.
-    ///
-    /// This method should _always_ be called during catalog creation _before_
-    /// any other operations that depend on system configuration values.
-    ///
-    /// Configuration is loaded in the following order:
-    ///
-    /// 1. Load parameters from the configuration persisted in the catalog
-    ///    storage backend.
-    /// 2. Set defaults from configuration passed in the provided
-    ///    `system_parameter_defaults` map.
-    /// 3. Overwrite and persist selected parameter values from the
-    ///    configuration that can be pulled from the provided
-    ///    `system_parameter_sync_config` (if present).
-    ///
-    /// # Errors
-    #[tracing::instrument(level = "info", skip_all)]
-    async fn load_system_configuration(
-        &mut self,
-        system_parameter_defaults: BTreeMap<String, String>,
-        system_parameter_sync_config: Option<SystemParameterSyncConfig>,
-        boot_ts: mz_repr::Timestamp,
-    ) -> Result<(), AdapterError> {
-        let system_config = self.storage().await.get_system_configurations().await?;
-
-        for (name, value) in &system_parameter_defaults {
-            match self
-                .state
-                .set_system_configuration_default(name, VarInput::Flat(value))
-            {
-                Ok(_) => (),
-                Err(AdapterError::VarError(VarError::UnknownParameter(name))) => {
-                    warn!(%name, "cannot load unknown system parameter from stash");
-                }
-                Err(e) => return Err(e),
-            };
-        }
-        for mz_catalog::SystemConfiguration { name, value } in system_config {
-            match self
-                .state
-                .insert_system_configuration(&name, VarInput::Flat(&value))
-            {
-                Ok(_) => (),
-                Err(AdapterError::VarError(VarError::UnknownParameter(name))) => {
-                    warn!(%name, "cannot load unknown system parameter from stash");
-                }
-                Err(e) => return Err(e),
-            };
-        }
-        if let Some(system_parameter_sync_config) = system_parameter_sync_config {
-            if self.storage().await.is_read_only() {
-                tracing::info!("parameter sync on boot: skipping sync as catalog is read-only");
-            } else if !self.state.system_config().config_has_synced_once() {
-                tracing::info!("parameter sync on boot: start sync");
-
-                // We intentionally block initial startup, potentially forever,
-                // on initializing LaunchDarkly. This may seem scary, but the
-                // alternative is even scarier. Over time, we expect that the
-                // compiled-in default values for the system parameters will
-                // drift substantially from the defaults configured in
-                // LaunchDarkly, to the point that starting an environment
-                // without loading the latest values from LaunchDarkly will
-                // result in running an untested configuration.
-                //
-                // Note this only applies during initial startup. Restarting
-                // after we've synced once doesn't block on LaunchDarkly, as it
-                // seems reasonable to assume that the last-synced configuration
-                // was valid enough.
-                //
-                // This philosophy appears to provide a good balance between not
-                // running untested configurations in production while also not
-                // making LaunchDarkly a "tier 1" dependency for existing
-                // environments.
-                //
-                // If this proves to be an issue, we could seek to address the
-                // configuration drift in a different way--for example, by
-                // writing a script that runs in CI nightly and checks for
-                // deviation between the compiled Rust code and LaunchDarkly.
-                //
-                // If it is absolutely necessary to bring up a new environment
-                // while LaunchDarkly is down, the following manual mitigation
-                // can be performed:
-                //
-                //    1. Edit the environmentd startup parameters to omit the
-                //       LaunchDarkly configuration.
-                //    2. Boot environmentd.
-                //    3. Run `ALTER SYSTEM config_has_synced_once = true`.
-                //    4. Adjust any other parameters as necessary to avoid
-                //       running a nonstandard configuration in production.
-                //    5. Edit the environmentd startup parameters to restore the
-                //       LaunchDarkly configuration, for when LaunchDarkly comes
-                //       back online.
-                //    6. Reboot environmentd.
-
-                let mut params = SynchronizedParameters::new(self.state.system_config().clone());
-                let frontend = SystemParameterFrontend::from(&system_parameter_sync_config).await?;
-                frontend.pull(&mut params);
-                let ops = params
-                    .modified()
-                    .into_iter()
-                    .map(|param| {
-                        let name = param.name;
-                        let value = param.value;
-                        tracing::debug!(name, value, "sync parameter");
-                        Op::UpdateSystemConfiguration {
-                            name,
-                            value: OwnedVarInput::Flat(value),
-                        }
-                    })
-                    .chain(std::iter::once({
-                        let name = CONFIG_HAS_SYNCED_ONCE.name().to_string();
-                        let value = true.to_string();
-                        tracing::debug!(name, value, "sync parameter");
-                        Op::UpdateSystemConfiguration {
-                            name,
-                            value: OwnedVarInput::Flat(value),
-                        }
-                    }))
-                    .collect::<Vec<_>>();
-                self.transact(boot_ts, None, ops, |_| Ok(()))
-                    .await
-                    .unwrap_or_terminate("cannot fail to transact");
-                tracing::info!("parameter sync on boot: end sync");
-            } else {
-                tracing::info!("parameter sync on boot: skipping sync as config has synced once");
-            }
-        }
-        Ok(())
-    }
-
-    /// Loads built-in system types into the catalog.
-    ///
-    /// Built-in types sometimes have references to other built-in types, and sometimes these
-    /// references are circular. This makes loading built-in types more complicated than other
-    /// built-in objects, and requires us to make multiple passes over the types to correctly
-    /// resolve all references.
-    #[tracing::instrument(level = "info", skip_all)]
-    async fn load_builtin_types(&mut self) -> Result<(), Error> {
-        let persisted_builtin_ids: BTreeMap<_, _> = self
-            .storage()
-            .await
-            .get_system_items()
-            .await?
-            .into_iter()
-            .map(|mapping| (mapping.description, mapping.unique_identifier))
-            .collect();
-
-        let AllocatedBuiltinSystemIds {
-            all_builtins,
-            new_builtins,
-            migrated_builtins,
-        } = self
-            .allocate_system_ids(BUILTINS::types().collect(), |typ| {
-                persisted_builtin_ids
-                    .get(&SystemObjectDescription {
-                        schema_name: typ.schema.to_string(),
-                        object_type: CatalogItemType::Type,
-                        object_name: typ.name.to_string(),
-                    })
-                    .cloned()
-            })
-            .await?;
-        assert!(migrated_builtins.is_empty(), "types cannot be migrated");
-        let name_to_id_map: BTreeMap<&str, GlobalId> = all_builtins
-            .into_iter()
-            .map(|(typ, id)| (typ.name, id))
-            .collect();
-
-        // Replace named references with id references
-        let mut builtin_types: Vec<_> = BUILTINS::types()
-            .map(|typ| Self::resolve_builtin_type(typ, &name_to_id_map))
-            .collect();
-
-        // Resolve array_id for types
-        let mut element_id_to_array_id = BTreeMap::new();
-        for typ in &builtin_types {
-            match &typ.details.typ {
-                CatalogType::Array { element_reference } => {
-                    let array_id = name_to_id_map[typ.name];
-                    element_id_to_array_id.insert(*element_reference, array_id);
-                }
-                _ => {}
-            }
-        }
-        let pg_catalog_schema_id = self.state.get_pg_catalog_schema_id().clone();
-        for typ in &mut builtin_types {
-            let element_id = name_to_id_map[typ.name];
-            typ.details.array_id = element_id_to_array_id.get(&element_id).map(|id| id.clone());
-        }
-
-        // Insert into catalog
-        for typ in builtin_types {
-            let element_id = name_to_id_map[typ.name];
-            self.state.insert_item(
-                element_id,
-                typ.oid,
-                QualifiedItemName {
-                    qualifiers: ItemQualifiers {
-                        database_spec: ResolvedDatabaseSpecifier::Ambient,
-                        schema_spec: SchemaSpecifier::Id(pg_catalog_schema_id),
-                    },
-                    item: typ.name.to_owned(),
-                },
-                CatalogItem::Type(Type {
-                    create_sql: format!("CREATE TYPE {}", typ.name),
-                    details: typ.details.clone(),
-                    resolved_ids: ResolvedIds(BTreeSet::new()),
-                }),
-                MZ_SYSTEM_ROLE_ID,
-                PrivilegeMap::from_mz_acl_items(vec![
-                    rbac::default_builtin_object_privilege(mz_sql::catalog::ObjectType::Type),
-                    rbac::owner_privilege(mz_sql::catalog::ObjectType::Type, MZ_SYSTEM_ROLE_ID),
-                ]),
-            );
-        }
-
-        let new_system_id_mappings = new_builtins
-            .iter()
-            .map(|(typ, id)| SystemObjectMapping {
-                description: SystemObjectDescription {
-                    schema_name: typ.schema.to_string(),
-                    object_type: CatalogItemType::Type,
-                    object_name: typ.name.to_string(),
-                },
-                unique_identifier: SystemObjectUniqueIdentifier {
-                    id: *id,
-                    fingerprint: typ.fingerprint(),
-                },
-            })
-            .collect();
-        self.storage()
-            .await
-            .set_system_items(new_system_id_mappings)
-            .await?;
-
-        Ok(())
-    }
-
     fn resolve_builtin_type(
         builtin: &BuiltinType<NameReference>,
         name_to_id_map: &BTreeMap<&str, GlobalId>,
@@ -3101,399 +414,11 @@ impl Catalog {
         }
     }
 
-    /// The objects in the catalog form one or more DAGs (directed acyclic graph) via object
-    /// dependencies. To migrate a builtin object we must drop that object along with all of its
-    /// descendants, and then recreate that object along with all of its descendants using new
-    /// GlobalId`s. To achieve this we perform a DFS (depth first search) on the catalog items
-    /// starting with the nodes that correspond to builtin objects that have changed schemas.
-    ///
-    /// Objects need to be dropped starting from the leafs of the DAG going up towards the roots,
-    /// and they need to be recreated starting at the roots of the DAG and going towards the leafs.
-    pub async fn generate_builtin_migration_metadata(
-        &self,
-        migrated_ids: Vec<GlobalId>,
-        id_fingerprint_map: BTreeMap<GlobalId, String>,
-    ) -> Result<BuiltinMigrationMetadata, Error> {
-        // First obtain a topological sorting of all migrated objects and their children.
-        let mut visited_set = BTreeSet::new();
-        let mut topological_sort = Vec::new();
-        for id in migrated_ids {
-            if !visited_set.contains(&id) {
-                let migrated_topological_sort = self.topological_sort(id, &mut visited_set);
-                topological_sort.extend(migrated_topological_sort);
-            }
-        }
-        topological_sort.reverse();
-
-        // Then process all objects in sorted order.
-        let mut migration_metadata = BuiltinMigrationMetadata::new();
-        let mut ancestor_ids = BTreeMap::new();
-        let mut migrated_log_ids = BTreeMap::new();
-        let log_name_map: BTreeMap<_, _> = BUILTINS::logs()
-            .map(|log| (log.variant.clone(), log.name))
-            .collect();
-        for entry in topological_sort {
-            let id = entry.id();
-
-            let new_id = match id {
-                GlobalId::System(_) => self
-                    .storage()
-                    .await
-                    .allocate_system_ids(1)
-                    .await?
-                    .into_element(),
-                GlobalId::User(_) => self.storage().await.allocate_user_id().await?,
-                _ => unreachable!("can't migrate id: {id}"),
-            };
-
-            let name = self.resolve_full_name(entry.name(), None);
-            info!("migrating {name} from {id} to {new_id}");
-
-            // Generate value to update fingerprint and global ID persisted mapping for system objects.
-            // Not every system object has a fingerprint, like introspection source indexes.
-            if let Some(fingerprint) = id_fingerprint_map.get(&id) {
-                assert!(
-                    id.is_system(),
-                    "id_fingerprint_map should only contain builtin objects"
-                );
-                let schema_name = self
-                    .get_schema(
-                        &entry.name.qualifiers.database_spec,
-                        &entry.name.qualifiers.schema_spec,
-                        entry.conn_id().unwrap_or(&SYSTEM_CONN_ID),
-                    )
-                    .name
-                    .schema
-                    .as_str();
-                migration_metadata.migrated_system_object_mappings.insert(
-                    id,
-                    SystemObjectMapping {
-                        description: SystemObjectDescription {
-                            schema_name: schema_name.to_string(),
-                            object_type: entry.item_type(),
-                            object_name: entry.name.item.clone(),
-                        },
-                        unique_identifier: SystemObjectUniqueIdentifier {
-                            id: new_id,
-                            fingerprint: fingerprint.clone(),
-                        },
-                    },
-                );
-            }
-
-            ancestor_ids.insert(id, new_id);
-
-            // Push drop commands.
-            match entry.item() {
-                CatalogItem::Table(_) | CatalogItem::Source(_) => {
-                    migration_metadata.previous_source_ids.push(id)
-                }
-                CatalogItem::Sink(_) => migration_metadata.previous_sink_ids.push(id),
-                CatalogItem::MaterializedView(_) => {
-                    migration_metadata.previous_materialized_view_ids.push(id)
-                }
-                CatalogItem::Log(log) => {
-                    migrated_log_ids.insert(id, log.variant.clone());
-                }
-                CatalogItem::Index(index) => {
-                    if id.is_system() {
-                        if let Some(variant) = migrated_log_ids.get(&index.on) {
-                            migration_metadata
-                                .introspection_source_index_updates
-                                .entry(index.cluster_id)
-                                .or_default()
-                                .push((
-                                    variant.clone(),
-                                    log_name_map
-                                        .get(variant)
-                                        .expect("all variants have a name")
-                                        .to_string(),
-                                    new_id,
-                                ));
-                        }
-                    }
-                }
-                CatalogItem::View(_) => {
-                    // Views don't have any external objects to drop.
-                }
-                CatalogItem::Type(_)
-                | CatalogItem::Func(_)
-                | CatalogItem::Secret(_)
-                | CatalogItem::Connection(_) => unreachable!(
-                    "impossible to migrate schema for builtin {}",
-                    entry.item().typ()
-                ),
-            }
-            if id.is_user() {
-                migration_metadata.user_drop_ops.push(id);
-            }
-            migration_metadata.all_drop_ops.push(id);
-
-            // Push create commands.
-            let name = entry.name.clone();
-            if id.is_user() {
-                let schema_id = name.qualifiers.schema_spec.clone().into();
-                migration_metadata
-                    .user_create_ops
-                    .push((new_id, schema_id, name.item.clone()));
-            }
-            let item_rebuilder = CatalogItemRebuilder::new(entry, new_id, &ancestor_ids);
-            migration_metadata.all_create_ops.push((
-                new_id,
-                entry.oid,
-                name,
-                entry.owner_id,
-                entry.privileges.clone(),
-                item_rebuilder,
-            ));
-        }
-
-        // Reverse drop commands.
-        migration_metadata.previous_sink_ids.reverse();
-        migration_metadata.previous_materialized_view_ids.reverse();
-        migration_metadata.previous_source_ids.reverse();
-        migration_metadata.all_drop_ops.reverse();
-        migration_metadata.user_drop_ops.reverse();
-
-        Ok(migration_metadata)
-    }
-
-    fn topological_sort(
-        &self,
-        id: GlobalId,
-        visited_set: &mut BTreeSet<GlobalId>,
-    ) -> Vec<&CatalogEntry> {
-        let mut topological_sort = Vec::new();
-        visited_set.insert(id);
-        let entry = self.get_entry(&id);
-        for dependant in &entry.used_by {
-            if !visited_set.contains(dependant) {
-                let child_topological_sort = self.topological_sort(*dependant, visited_set);
-                topological_sort.extend(child_topological_sort);
-            }
-        }
-        topological_sort.push(entry);
-        topological_sort
-    }
-
-    pub fn apply_in_memory_builtin_migration(
-        &mut self,
-        migration_metadata: &mut BuiltinMigrationMetadata,
-    ) -> Result<(), Error> {
-        assert_eq!(
-            migration_metadata.all_drop_ops.len(),
-            migration_metadata.all_create_ops.len(),
-            "we should be re-creating every dropped object"
-        );
-        for id in migration_metadata.all_drop_ops.drain(..) {
-            self.state.drop_item(id);
-            self.drop_plans_and_metainfos(id);
-        }
-        for (id, oid, name, owner_id, privileges, item_rebuilder) in
-            migration_metadata.all_create_ops.drain(..)
-        {
-            let item = item_rebuilder.build(self);
-            self.state
-                .insert_item(id, oid, name, item, owner_id, privileges);
-        }
-        for (cluster_id, updates) in &migration_metadata.introspection_source_index_updates {
-            let log_indexes = &mut self
-                .state
-                .clusters_by_id
-                .get_mut(cluster_id)
-                .unwrap_or_else(|| panic!("invalid cluster {cluster_id}"))
-                .log_indexes;
-            for (variant, _name, new_id) in updates {
-                log_indexes.remove(variant);
-                log_indexes.insert(variant.clone(), new_id.clone());
-            }
-        }
-
-        Ok(())
-    }
-
-    #[tracing::instrument(level = "info", skip_all)]
-    pub async fn apply_persisted_builtin_migration(
-        &self,
-        migration_metadata: &mut BuiltinMigrationMetadata,
-    ) -> Result<(), Error> {
-        let mut storage = self.storage().await;
-        let mut tx = storage.transaction().await?;
-        tx.remove_items(migration_metadata.user_drop_ops.drain(..).collect())?;
-        for (id, schema_id, name) in migration_metadata.user_create_ops.drain(..) {
-            let entry = self.get_entry(&id);
-            let item = entry.item();
-            let serialized_item = item.to_serialized();
-            tx.insert_item(
-                id,
-                schema_id,
-                &name,
-                serialized_item,
-                entry.owner_id,
-                entry.privileges().all_values_owned().collect(),
-            )?;
-        }
-        tx.update_system_object_mappings(std::mem::take(
-            &mut migration_metadata.migrated_system_object_mappings,
-        ))?;
-        tx.update_introspection_source_index_gids(
-            std::mem::take(&mut migration_metadata.introspection_source_index_updates)
-                .into_iter()
-                .map(|(cluster_id, updates)| {
-                    (
-                        cluster_id,
-                        updates
-                            .into_iter()
-                            .map(|(_variant, name, index_id)| (name, index_id)),
-                    )
-                }),
-        )?;
-
-        tx.commit().await?;
-
-        Ok(())
-    }
-
     /// Returns the catalog's transient revision, which starts at 1 and is
     /// incremented on every change. This is not persisted to disk, and will
     /// restart on every load.
     pub fn transient_revision(&self) -> u64 {
         self.transient_revision
-    }
-
-    /// Takes a catalog which only has items in its on-disk storage ("unloaded")
-    /// and cannot yet resolve names, and returns a catalog loaded with those
-    /// items.
-    ///
-    /// This function requires transactions to support loading a catalog with
-    /// the transaction's currently in-flight updates to existing catalog
-    /// objects, which is necessary for at least one catalog migration.
-    ///
-    /// TODO(justin): it might be nice if these were two different types.
-    #[tracing::instrument(level = "info", skip_all)]
-    pub fn load_catalog_items<'a>(
-        tx: &mut mz_catalog::Transaction<'a>,
-        c: &Catalog,
-    ) -> Result<Catalog, Error> {
-        let mut c = c.clone();
-        let mut awaiting_id_dependencies: BTreeMap<GlobalId, Vec<_>> = BTreeMap::new();
-        let mut awaiting_name_dependencies: BTreeMap<String, Vec<_>> = BTreeMap::new();
-        let mut items: VecDeque<_> = tx.loaded_items().into_iter().collect();
-        while let Some(item) = items.pop_front() {
-            let d_c = item.create_sql.clone();
-            // TODO(benesch): a better way of detecting when a view has depended
-            // upon a non-existent logging view. This is fine for now because
-            // the only goal is to produce a nicer error message; we'll bail out
-            // safely even if the error message we're sniffing out changes.
-            static LOGGING_ERROR: Lazy<Regex> =
-                Lazy::new(|| Regex::new("mz_catalog.[^']*").expect("valid regex"));
-
-            let catalog_item = match c.deserialize_item(item.id, d_c) {
-                Ok(item) => item,
-                Err(AdapterError::Catalog(Error {
-                    kind: ErrorKind::Sql(SqlCatalogError::UnknownItem(name)),
-                })) if LOGGING_ERROR.is_match(&name.to_string()) => {
-                    return Err(Error::new(ErrorKind::UnsatisfiableLoggingDependency {
-                        depender_name: name,
-                    }));
-                }
-                // If we were missing a dependency, wait for it to be added.
-                Err(AdapterError::PlanError(plan::PlanError::InvalidId(missing_dep))) => {
-                    awaiting_id_dependencies
-                        .entry(missing_dep)
-                        .or_default()
-                        .push(item);
-                    continue;
-                }
-                // If we were missing a dependency, wait for it to be added.
-                Err(AdapterError::PlanError(plan::PlanError::Catalog(
-                    SqlCatalogError::UnknownItem(missing_dep),
-                ))) => {
-                    match GlobalId::from_str(&missing_dep) {
-                        Ok(id) => {
-                            awaiting_id_dependencies.entry(id).or_default().push(item);
-                        }
-                        Err(_) => {
-                            awaiting_name_dependencies
-                                .entry(missing_dep)
-                                .or_default()
-                                .push(item);
-                        }
-                    }
-                    continue;
-                }
-                Err(e) => {
-                    let name = c.resolve_full_name(&item.name, None);
-                    return Err(Error::new(ErrorKind::Corruption {
-                        detail: format!("failed to deserialize item {} ({}): {}", item.id, name, e),
-                    }));
-                }
-            };
-            let oid = c.allocate_oid()?;
-
-            // Enqueue any items waiting on this dependency.
-            if let Some(dependent_items) = awaiting_id_dependencies.remove(&item.id) {
-                items.extend(dependent_items);
-            }
-            let full_name = c.resolve_full_name(&item.name, None);
-            if let Some(dependent_items) = awaiting_name_dependencies.remove(&full_name.to_string())
-            {
-                items.extend(dependent_items);
-            }
-
-            c.state.insert_item(
-                item.id,
-                oid,
-                item.name,
-                catalog_item,
-                item.owner_id,
-                PrivilegeMap::from_mz_acl_items(item.privileges),
-            );
-        }
-
-        // Error on any unsatisfied dependencies.
-        if let Some((missing_dep, mut dependents)) = awaiting_id_dependencies.into_iter().next() {
-            let mz_catalog::Item {
-                id,
-                name,
-                create_sql: _,
-                owner_id: _,
-                privileges: _,
-            } = dependents.remove(0);
-            let name = c.resolve_full_name(&name, None);
-            return Err(Error::new(ErrorKind::Corruption {
-                detail: format!(
-                    "failed to deserialize item {} ({}): {}",
-                    id,
-                    name,
-                    AdapterError::PlanError(plan::PlanError::InvalidId(missing_dep))
-                ),
-            }));
-        }
-
-        if let Some((missing_dep, mut dependents)) = awaiting_name_dependencies.into_iter().next() {
-            let mz_catalog::Item {
-                id,
-                name,
-                create_sql: _,
-                owner_id: _,
-                privileges: _,
-            } = dependents.remove(0);
-            let name = c.resolve_full_name(&name, None);
-            return Err(Error::new(ErrorKind::Corruption {
-                detail: format!(
-                    "failed to deserialize item {} ({}): {}",
-                    id,
-                    name,
-                    AdapterError::Catalog(Error {
-                        kind: ErrorKind::Sql(SqlCatalogError::UnknownItem(missing_dep))
-                    })
-                ),
-            }));
-        }
-
-        c.transient_revision = 1;
-        Ok(c)
     }
 
     /// Creates a debug catalog from a debug stash based on the current
@@ -3686,63 +611,6 @@ impl Catalog {
 
     async fn storage<'a>(&'a self) -> MutexGuard<'a, Box<dyn mz_catalog::DurableCatalogState>> {
         self.storage.lock().await
-    }
-
-    /// Allocate new system ids for any new builtin objects and looks up existing system ids for
-    /// existing builtin objects
-    async fn allocate_system_ids<T, F>(
-        &self,
-        builtins: Vec<T>,
-        builtin_lookup: F,
-    ) -> Result<AllocatedBuiltinSystemIds<T>, Error>
-    where
-        T: Copy + Fingerprint,
-        F: Fn(&T) -> Option<SystemObjectUniqueIdentifier>,
-    {
-        let new_builtin_amount = builtins
-            .iter()
-            .filter(|builtin| builtin_lookup(builtin).is_none())
-            .count();
-
-        let mut global_ids = self
-            .storage()
-            .await
-            .allocate_system_ids(
-                new_builtin_amount
-                    .try_into()
-                    .expect("builtins should fit into u64"),
-            )
-            .await?
-            .into_iter();
-
-        let mut all_builtins = Vec::new();
-        let mut new_builtins = Vec::new();
-        let mut migrated_builtins = Vec::new();
-        for builtin in &builtins {
-            match builtin_lookup(builtin) {
-                Some(SystemObjectUniqueIdentifier {
-                    id,
-                    fingerprint: old_fingerprint,
-                }) => {
-                    all_builtins.push((*builtin, id));
-                    let new_fingerprint = builtin.fingerprint();
-                    if old_fingerprint != new_fingerprint {
-                        migrated_builtins.push(id);
-                    }
-                }
-                None => {
-                    let id = global_ids.next().expect("not enough global IDs");
-                    all_builtins.push((*builtin, id));
-                    new_builtins.push((*builtin, id));
-                }
-            }
-        }
-
-        Ok(AllocatedBuiltinSystemIds {
-            all_builtins,
-            new_builtins,
-            migrated_builtins,
-        })
     }
 
     pub async fn allocate_user_id(&self) -> Result<GlobalId, Error> {
@@ -4370,7 +1238,7 @@ impl Catalog {
             .iter()
             .filter_map(|id| {
                 let entry = self.get_entry(id);
-                match entry.item.conn_id() {
+                match entry.item().conn_id() {
                     Some(conn_id) => Some((conn_id, entry.name().item.clone())),
                     None => None,
                 }
@@ -5501,7 +2369,7 @@ impl Catalog {
                             }
 
                             builtin_table_updates.extend(state.pack_item_update(id, -1));
-                            if Self::should_audit_log_item(&entry.item) {
+                            if Self::should_audit_log_item(entry.item()) {
                                 state.add_to_audit_log(
                                     oracle_write_ts,
                                     session,
@@ -5513,7 +2381,7 @@ impl Catalog {
                                     EventDetails::IdFullNameV1(IdFullNameV1 {
                                         id: id.to_string(),
                                         name: Self::full_name_detail(&state.resolve_full_name(
-                                            &entry.name,
+                                            entry.name(),
                                             session.map(|session| session.conn_id()),
                                         )),
                                     }),
@@ -5899,7 +2767,7 @@ impl Catalog {
                         old_name: Self::full_name_detail(&current_full_name),
                         new_name: Self::full_name_detail(&to_full_name),
                     });
-                    if Self::should_audit_log_item(&entry.item) {
+                    if Self::should_audit_log_item(entry.item()) {
                         state.add_to_audit_log(
                             oracle_write_ts,
                             session,
@@ -5916,7 +2784,7 @@ impl Catalog {
                     let mut new_entry = entry.clone();
                     new_entry.name.item = to_name.clone();
                     new_entry.item = entry
-                        .item
+                        .item()
                         .rename_item_refs(
                             current_full_name.clone(),
                             to_full_name.item.clone(),
@@ -5925,10 +2793,10 @@ impl Catalog {
                         .map_err(|e| {
                             Error::new(ErrorKind::from(AmbiguousRename {
                                 depender: state
-                                    .resolve_full_name(&entry.name, entry.conn_id())
+                                    .resolve_full_name(entry.name(), entry.conn_id())
                                     .to_string(),
                                 dependee: state
-                                    .resolve_full_name(&entry.name, entry.conn_id())
+                                    .resolve_full_name(entry.name(), entry.conn_id())
                                     .to_string(),
                                 message: e,
                             }))
@@ -5938,7 +2806,7 @@ impl Catalog {
                         let dependent_item = state.get_entry(id);
                         let mut to_entry = dependent_item.clone();
                         to_entry.item = dependent_item
-                            .item
+                            .item()
                             .rename_item_refs(
                                 current_full_name.clone(),
                                 to_full_name.item.clone(),
@@ -5948,12 +2816,12 @@ impl Catalog {
                                 Error::new(ErrorKind::from(AmbiguousRename {
                                     depender: state
                                         .resolve_full_name(
-                                            &dependent_item.name,
+                                            dependent_item.name(),
                                             dependent_item.conn_id(),
                                         )
                                         .to_string(),
                                     dependee: state
-                                        .resolve_full_name(&entry.name, entry.conn_id())
+                                        .resolve_full_name(entry.name(), entry.conn_id())
                                         .to_string(),
                                     message: e,
                                 }))
@@ -6254,7 +3122,7 @@ impl Catalog {
                     info!(
                         "update {} {} ({})",
                         old_entry.item_type(),
-                        state.resolve_full_name(&old_entry.name, old_entry.conn_id()),
+                        state.resolve_full_name(old_entry.name(), old_entry.conn_id()),
                         id
                     );
                     let mut new_entry = old_entry;
@@ -6280,7 +3148,7 @@ impl Catalog {
         info!(
             "update {} {} ({})",
             old_entry.item_type(),
-            state.resolve_full_name(&old_entry.name, old_entry.conn_id()),
+            state.resolve_full_name(old_entry.name(), old_entry.conn_id()),
             id
         );
 
@@ -6328,11 +3196,11 @@ impl Catalog {
 
         for u in new_deps {
             match state.entry_by_id.get_mut(&u) {
-                Some(metadata) => metadata.used_by.push(new_entry.id),
+                Some(metadata) => metadata.used_by.push(new_entry.id()),
                 None => panic!(
                     "Catalog: missing dependent catalog item {} while updating {}",
                     &u,
-                    state.resolve_full_name(&new_entry.name, new_entry.conn_id())
+                    state.resolve_full_name(new_entry.name(), new_entry.conn_id())
                 ),
             }
         }
@@ -6669,32 +3537,32 @@ impl Catalog {
 
     pub fn user_connections(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_connection() && entry.id.is_user())
+            .filter(|entry| entry.is_connection() && entry.id().is_user())
     }
 
     pub fn user_tables(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_table() && entry.id.is_user())
+            .filter(|entry| entry.is_table() && entry.id().is_user())
     }
 
     pub fn user_sources(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_source() && entry.id.is_user())
+            .filter(|entry| entry.is_source() && entry.id().is_user())
     }
 
     pub fn user_sinks(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_sink() && entry.id.is_user())
+            .filter(|entry| entry.is_sink() && entry.id().is_user())
     }
 
     pub fn user_materialized_views(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_materialized_view() && entry.id.is_user())
+            .filter(|entry| entry.is_materialized_view() && entry.id().is_user())
     }
 
     pub fn user_secrets(&self) -> impl Iterator<Item = &CatalogEntry> {
         self.entries()
-            .filter(|entry| entry.is_secret() && entry.id.is_user())
+            .filter(|entry| entry.is_secret() && entry.id().is_user())
     }
 
     pub fn clusters(&self) -> impl Iterator<Item = &Cluster> {
@@ -7045,87 +3913,6 @@ pub enum Op {
     },
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize, PartialOrd, PartialEq, Eq, Ord)]
-pub struct ClusterConfig {
-    pub variant: ClusterVariant,
-}
-
-impl From<ClusterConfig> for mz_catalog::ClusterConfig {
-    fn from(config: ClusterConfig) -> Self {
-        Self {
-            variant: config.variant.into(),
-        }
-    }
-}
-
-impl From<mz_catalog::ClusterConfig> for ClusterConfig {
-    fn from(config: mz_catalog::ClusterConfig) -> Self {
-        Self {
-            variant: config.variant.into(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, PartialOrd, PartialEq, Eq, Ord)]
-pub struct ClusterVariantManaged {
-    pub size: String,
-    pub availability_zones: Vec<String>,
-    pub logging: ReplicaLogging,
-    pub idle_arrangement_merge_effort: Option<u32>,
-    pub replication_factor: u32,
-    pub disk: bool,
-}
-
-impl From<ClusterVariantManaged> for mz_catalog::ClusterVariantManaged {
-    fn from(managed: ClusterVariantManaged) -> Self {
-        Self {
-            size: managed.size,
-            availability_zones: managed.availability_zones,
-            logging: managed.logging,
-            idle_arrangement_merge_effort: managed.idle_arrangement_merge_effort,
-            replication_factor: managed.replication_factor,
-            disk: managed.disk,
-        }
-    }
-}
-
-impl From<mz_catalog::ClusterVariantManaged> for ClusterVariantManaged {
-    fn from(managed: mz_catalog::ClusterVariantManaged) -> Self {
-        Self {
-            size: managed.size,
-            availability_zones: managed.availability_zones,
-            logging: managed.logging,
-            idle_arrangement_merge_effort: managed.idle_arrangement_merge_effort,
-            replication_factor: managed.replication_factor,
-            disk: managed.disk,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize, PartialOrd, PartialEq, Eq, Ord)]
-pub enum ClusterVariant {
-    Managed(ClusterVariantManaged),
-    Unmanaged,
-}
-
-impl From<ClusterVariant> for mz_catalog::ClusterVariant {
-    fn from(variant: ClusterVariant) -> Self {
-        match variant {
-            ClusterVariant::Managed(managed) => Self::Managed(managed.into()),
-            ClusterVariant::Unmanaged => Self::Unmanaged,
-        }
-    }
-}
-
-impl From<mz_catalog::ClusterVariant> for ClusterVariant {
-    fn from(variant: mz_catalog::ClusterVariant) -> Self {
-        match variant {
-            mz_catalog::ClusterVariant::Managed(managed) => Self::Managed(managed.into()),
-            mz_catalog::ClusterVariant::Unmanaged => Self::Unmanaged,
-        }
-    }
-}
-
 impl ConnCatalog<'_> {
     fn resolve_item_name(
         &self,
@@ -7235,8 +4022,8 @@ impl ExprHumanizer for ConnCatalog<'_> {
             .try_map(|entry| {
                 Ok::<_, SqlCatalogError>(
                     entry
-                        .item
-                        .desc(&self.resolve_full_name(&entry.name))?
+                        .item()
+                        .desc(&self.resolve_full_name(entry.name()))?
                         .iter_names()
                         .cloned()
                         .map(|col_name| col_name.to_string())
@@ -7646,266 +4433,6 @@ impl SessionCatalog for ConnCatalog<'_> {
     }
 }
 
-impl mz_sql::catalog::CatalogDatabase for Database {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn id(&self) -> DatabaseId {
-        self.id
-    }
-
-    fn has_schemas(&self) -> bool {
-        !self.schemas_by_name.is_empty()
-    }
-
-    fn schema_ids(&self) -> &BTreeMap<String, SchemaId> {
-        &self.schemas_by_name
-    }
-
-    // `as` is ok to use to cast to a trait object.
-    #[allow(clippy::as_conversions)]
-    fn schemas(&self) -> Vec<&dyn CatalogSchema> {
-        self.schemas_by_id
-            .values()
-            .map(|schema| schema as &dyn CatalogSchema)
-            .collect()
-    }
-
-    fn owner_id(&self) -> RoleId {
-        self.owner_id
-    }
-
-    fn privileges(&self) -> &PrivilegeMap {
-        &self.privileges
-    }
-}
-
-impl mz_sql::catalog::CatalogSchema for Schema {
-    fn database(&self) -> &ResolvedDatabaseSpecifier {
-        &self.name.database
-    }
-
-    fn name(&self) -> &QualifiedSchemaName {
-        &self.name
-    }
-
-    fn id(&self) -> &SchemaSpecifier {
-        &self.id
-    }
-
-    fn has_items(&self) -> bool {
-        !self.items.is_empty()
-    }
-
-    fn item_ids(&self) -> &BTreeMap<String, GlobalId> {
-        &self.items
-    }
-
-    fn owner_id(&self) -> RoleId {
-        self.owner_id
-    }
-
-    fn privileges(&self) -> &PrivilegeMap {
-        &self.privileges
-    }
-}
-
-impl mz_sql::catalog::CatalogRole for Role {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn id(&self) -> RoleId {
-        self.id
-    }
-
-    fn membership(&self) -> &BTreeMap<RoleId, RoleId> {
-        &self.membership.map
-    }
-
-    fn attributes(&self) -> &RoleAttributes {
-        &self.attributes
-    }
-
-    fn vars(&self) -> &BTreeMap<String, OwnedVarInput> {
-        &self.vars.map
-    }
-}
-
-impl mz_sql::catalog::CatalogCluster<'_> for Cluster {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn id(&self) -> ClusterId {
-        self.id
-    }
-
-    fn linked_object_id(&self) -> Option<GlobalId> {
-        self.linked_object_id
-    }
-
-    fn bound_objects(&self) -> &BTreeSet<GlobalId> {
-        &self.bound_objects
-    }
-
-    fn replica_ids(&self) -> &BTreeMap<String, ReplicaId> {
-        &self.replica_id_by_name_
-    }
-
-    // `as` is ok to use to cast to a trait object.
-    #[allow(clippy::as_conversions)]
-    fn replicas(&self) -> Vec<&dyn CatalogClusterReplica> {
-        self.replicas()
-            .map(|replica| replica as &dyn CatalogClusterReplica)
-            .collect()
-    }
-
-    fn replica(&self, id: ReplicaId) -> &dyn CatalogClusterReplica {
-        self.replica(id).expect("catalog out of sync")
-    }
-
-    fn owner_id(&self) -> RoleId {
-        self.owner_id
-    }
-
-    fn privileges(&self) -> &PrivilegeMap {
-        &self.privileges
-    }
-
-    fn is_managed(&self) -> bool {
-        self.is_managed()
-    }
-}
-
-impl mz_sql::catalog::CatalogClusterReplica<'_> for ClusterReplica {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn cluster_id(&self) -> ClusterId {
-        self.cluster_id
-    }
-
-    fn replica_id(&self) -> ReplicaId {
-        self.replica_id
-    }
-
-    fn owner_id(&self) -> RoleId {
-        self.owner_id
-    }
-}
-
-impl mz_sql::catalog::CatalogItem for CatalogEntry {
-    fn name(&self) -> &QualifiedItemName {
-        self.name()
-    }
-
-    fn id(&self) -> GlobalId {
-        self.id()
-    }
-
-    fn oid(&self) -> u32 {
-        self.oid()
-    }
-
-    fn desc(&self, name: &FullItemName) -> Result<Cow<RelationDesc>, SqlCatalogError> {
-        self.desc(name)
-    }
-
-    fn func(&self) -> Result<&'static mz_sql::func::Func, SqlCatalogError> {
-        self.func()
-    }
-
-    fn source_desc(&self) -> Result<Option<&SourceDesc<ReferencedConnection>>, SqlCatalogError> {
-        self.source_desc()
-    }
-
-    fn connection(
-        &self,
-    ) -> Result<&mz_storage_types::connections::Connection<ReferencedConnection>, SqlCatalogError>
-    {
-        Ok(&self.connection()?.connection)
-    }
-
-    fn create_sql(&self) -> &str {
-        match self.item() {
-            CatalogItem::Table(Table { create_sql, .. }) => create_sql,
-            CatalogItem::Source(Source { create_sql, .. }) => create_sql,
-            CatalogItem::Sink(Sink { create_sql, .. }) => create_sql,
-            CatalogItem::View(View { create_sql, .. }) => create_sql,
-            CatalogItem::MaterializedView(MaterializedView { create_sql, .. }) => create_sql,
-            CatalogItem::Index(Index { create_sql, .. }) => create_sql,
-            CatalogItem::Type(Type { create_sql, .. }) => create_sql,
-            CatalogItem::Secret(Secret { create_sql, .. }) => create_sql,
-            CatalogItem::Connection(Connection { create_sql, .. }) => create_sql,
-            CatalogItem::Func(_) => "<builtin>",
-            CatalogItem::Log(_) => "<builtin>",
-        }
-    }
-
-    fn item_type(&self) -> SqlCatalogItemType {
-        self.item().typ()
-    }
-
-    fn index_details(&self) -> Option<(&[MirScalarExpr], GlobalId)> {
-        if let CatalogItem::Index(Index { keys, on, .. }) = self.item() {
-            Some((keys, *on))
-        } else {
-            None
-        }
-    }
-
-    fn table_details(&self) -> Option<&[Expr<Aug>]> {
-        if let CatalogItem::Table(Table { defaults, .. }) = self.item() {
-            Some(defaults)
-        } else {
-            None
-        }
-    }
-
-    fn type_details(&self) -> Option<&CatalogTypeDetails<IdReference>> {
-        if let CatalogItem::Type(Type { details, .. }) = self.item() {
-            Some(details)
-        } else {
-            None
-        }
-    }
-
-    fn uses(&self) -> &ResolvedIds {
-        self.uses()
-    }
-
-    fn used_by(&self) -> &[GlobalId] {
-        self.used_by()
-    }
-
-    fn is_subsource(&self) -> bool {
-        self.is_subsource()
-    }
-
-    fn subsources(&self) -> BTreeSet<GlobalId> {
-        self.subsources()
-    }
-
-    fn progress_id(&self) -> Option<GlobalId> {
-        self.progress_id()
-    }
-
-    fn owner_id(&self) -> RoleId {
-        self.owner_id
-    }
-
-    fn privileges(&self) -> &PrivilegeMap {
-        &self.privileges
-    }
-
-    fn cluster_id(&self) -> Option<ClusterId> {
-        self.item().cluster_id()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::{BTreeMap, BTreeSet};
@@ -7919,20 +4446,20 @@ mod tests {
 
     use mz_catalog::builtin::{Builtin, BuiltinType, BUILTINS};
     use mz_controller_types::{ClusterId, ReplicaId};
-    use mz_expr::{MirRelationExpr, MirScalarExpr, OptimizedMirRelationExpr};
+    use mz_expr::MirScalarExpr;
     use mz_ore::collections::CollectionExt;
     use mz_ore::now::{to_datetime, NOW_ZERO, SYSTEM_TIME};
     use mz_ore::task;
     use mz_pgrepr::oid::{FIRST_MATERIALIZE_OID, FIRST_UNPINNED_OID};
-    use mz_repr::adt::mz_acl_item::{AclMode, MzAclItem, PrivilegeMap};
+    use mz_repr::adt::mz_acl_item::{AclMode, MzAclItem};
     use mz_repr::namespaces::{INFORMATION_SCHEMA, PG_CATALOG_SCHEMA};
     use mz_repr::role_id::RoleId;
-    use mz_repr::{Datum, GlobalId, RelationDesc, RelationType, RowArena, ScalarType, Timestamp};
-    use mz_sql::catalog::{CatalogDatabase, CatalogSchema, CatalogType, SessionCatalog};
+    use mz_repr::{Datum, GlobalId, RelationType, RowArena, ScalarType, Timestamp};
+    use mz_sql::catalog::{CatalogSchema, CatalogType, SessionCatalog};
     use mz_sql::func::{Func, FuncImpl, Operation, OP_IMPLS};
     use mz_sql::names::{
         self, DatabaseId, ItemQualifiers, ObjectId, PartialItemName, QualifiedItemName,
-        ResolvedDatabaseSpecifier, ResolvedIds, SchemaId, SchemaSpecifier, SystemObjectId,
+        ResolvedDatabaseSpecifier, SchemaId, SchemaSpecifier, SystemObjectId,
     };
     use mz_sql::plan::{
         CoercibleScalarExpr, ExprContext, HirScalarExpr, PlanContext, QueryContext, QueryLifetime,
@@ -7940,15 +4467,11 @@ mod tests {
     };
     use mz_sql::session::user::MZ_SYSTEM_ROLE_ID;
     use mz_sql::session::vars::VarInput;
-    use mz_sql::DEFAULT_SCHEMA;
-    use mz_sql_parser::ast::Expr;
     use mz_stash::DebugStashFactory;
 
-    use crate::catalog::{
-        Catalog, CatalogItem, Index, MaterializedView, Op, Table, SYSTEM_CONN_ID,
-    };
+    use crate::catalog::{Catalog, CatalogItem, Op, PrivilegeMap, SYSTEM_CONN_ID};
     use crate::coord::dataflows::{prep_scalar_expr, EvalTime, ExprPrepStyle};
-    use crate::session::{Session, DEFAULT_DATABASE_NAME};
+    use crate::session::Session;
 
     /// System sessions have an empty `search_path` so it's necessary to
     /// schema-qualify all referenced items.
@@ -8210,661 +4733,6 @@ mod tests {
             );
         })
         .await
-    }
-
-    #[mz_ore::test(tokio::test)]
-    #[cfg_attr(miri, ignore)] //  unsupported operation: can't call foreign function `TLS_client_method` on OS `linux`
-    async fn test_builtin_migration() {
-        enum ItemNamespace {
-            System,
-            User,
-        }
-
-        enum SimplifiedItem {
-            Table,
-            MaterializedView { referenced_names: Vec<String> },
-            Index { on: String },
-        }
-
-        struct SimplifiedCatalogEntry {
-            name: String,
-            namespace: ItemNamespace,
-            item: SimplifiedItem,
-        }
-
-        impl SimplifiedCatalogEntry {
-            // A lot of the fields here aren't actually used in the test so we can fill them in with dummy
-            // values.
-            fn to_catalog_item(
-                self,
-                id_mapping: &BTreeMap<String, GlobalId>,
-            ) -> (String, ItemNamespace, CatalogItem) {
-                let item = match self.item {
-                    SimplifiedItem::Table => CatalogItem::Table(Table {
-                        create_sql: "TODO".to_string(),
-                        desc: RelationDesc::empty()
-                            .with_column("a", ScalarType::Int32.nullable(true))
-                            .with_key(vec![0]),
-                        defaults: vec![Expr::null(); 1],
-                        conn_id: None,
-                        resolved_ids: ResolvedIds(BTreeSet::new()),
-                        custom_logical_compaction_window: None,
-                        is_retained_metrics_object: false,
-                    }),
-                    SimplifiedItem::MaterializedView { referenced_names } => {
-                        let table_list = referenced_names.iter().join(",");
-                        let resolved_ids = convert_name_vec_to_id_vec(referenced_names, id_mapping);
-                        CatalogItem::MaterializedView(MaterializedView {
-                            create_sql: format!(
-                                "CREATE MATERIALIZED VIEW mv AS SELECT * FROM {table_list}"
-                            ),
-                            optimized_expr: OptimizedMirRelationExpr(MirRelationExpr::Constant {
-                                rows: Ok(Vec::new()),
-                                typ: RelationType {
-                                    column_types: Vec::new(),
-                                    keys: Vec::new(),
-                                },
-                            }),
-                            desc: RelationDesc::empty()
-                                .with_column("a", ScalarType::Int32.nullable(true))
-                                .with_key(vec![0]),
-                            resolved_ids: ResolvedIds(BTreeSet::from_iter(resolved_ids)),
-                            cluster_id: ClusterId::User(1),
-                        })
-                    }
-                    SimplifiedItem::Index { on } => {
-                        let on_id = id_mapping[&on];
-                        CatalogItem::Index(Index {
-                            create_sql: format!("CREATE INDEX idx ON {on} (a)"),
-                            on: on_id,
-                            keys: Vec::new(),
-                            conn_id: None,
-                            resolved_ids: ResolvedIds(BTreeSet::from_iter([on_id])),
-                            cluster_id: ClusterId::User(1),
-                            custom_logical_compaction_window: None,
-                            is_retained_metrics_object: false,
-                        })
-                    }
-                };
-                (self.name, self.namespace, item)
-            }
-        }
-
-        struct BuiltinMigrationTestCase {
-            test_name: &'static str,
-            initial_state: Vec<SimplifiedCatalogEntry>,
-            migrated_names: Vec<String>,
-            expected_previous_sink_names: Vec<String>,
-            expected_previous_materialized_view_names: Vec<String>,
-            expected_previous_source_names: Vec<String>,
-            expected_all_drop_ops: Vec<String>,
-            expected_user_drop_ops: Vec<String>,
-            expected_all_create_ops: Vec<String>,
-            expected_user_create_ops: Vec<String>,
-            expected_migrated_system_object_mappings: Vec<String>,
-        }
-
-        async fn add_item(
-            catalog: &mut Catalog,
-            name: String,
-            item: CatalogItem,
-            item_namespace: ItemNamespace,
-        ) -> GlobalId {
-            let id = match item_namespace {
-                ItemNamespace::User => catalog
-                    .allocate_user_id()
-                    .await
-                    .expect("cannot fail to allocate user ids"),
-                ItemNamespace::System => catalog
-                    .allocate_system_id()
-                    .await
-                    .expect("cannot fail to allocate system ids"),
-            };
-            let oid = catalog
-                .allocate_oid()
-                .expect("cannot fail to allocate oids");
-            let database_id = catalog
-                .resolve_database(DEFAULT_DATABASE_NAME)
-                .expect("failed to resolve default database")
-                .id();
-            let database_spec = ResolvedDatabaseSpecifier::Id(database_id);
-            let schema_spec = catalog
-                .resolve_schema_in_database(&database_spec, DEFAULT_SCHEMA, &SYSTEM_CONN_ID)
-                .expect("failed to resolve default schemazs")
-                .id
-                .clone();
-            catalog
-                .transact(
-                    mz_repr::Timestamp::MIN,
-                    None,
-                    vec![Op::CreateItem {
-                        id,
-                        oid,
-                        name: QualifiedItemName {
-                            qualifiers: ItemQualifiers {
-                                database_spec,
-                                schema_spec,
-                            },
-                            item: name,
-                        },
-                        item,
-                        owner_id: MZ_SYSTEM_ROLE_ID,
-                    }],
-                    |_| Ok(()),
-                )
-                .await
-                .expect("failed to transact");
-            id
-        }
-
-        fn convert_name_vec_to_id_vec(
-            name_vec: Vec<String>,
-            id_lookup: &BTreeMap<String, GlobalId>,
-        ) -> Vec<GlobalId> {
-            name_vec.into_iter().map(|name| id_lookup[&name]).collect()
-        }
-
-        fn convert_id_vec_to_name_vec(
-            id_vec: Vec<GlobalId>,
-            name_lookup: &BTreeMap<GlobalId, String>,
-        ) -> Vec<String> {
-            id_vec
-                .into_iter()
-                .map(|id| name_lookup[&id].clone())
-                .collect()
-        }
-
-        let test_cases = vec![
-            BuiltinMigrationTestCase {
-                test_name: "no_migrations",
-                initial_state: vec![SimplifiedCatalogEntry {
-                    name: "s1".to_string(),
-                    namespace: ItemNamespace::System,
-                    item: SimplifiedItem::Table,
-                }],
-                migrated_names: vec![],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec![],
-                expected_previous_source_names: vec![],
-                expected_all_drop_ops: vec![],
-                expected_user_drop_ops: vec![],
-                expected_all_create_ops: vec![],
-                expected_user_create_ops: vec![],
-                expected_migrated_system_object_mappings: vec![],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "single_migrations",
-                initial_state: vec![SimplifiedCatalogEntry {
-                    name: "s1".to_string(),
-                    namespace: ItemNamespace::System,
-                    item: SimplifiedItem::Table,
-                }],
-                migrated_names: vec!["s1".to_string()],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec![],
-                expected_previous_source_names: vec!["s1".to_string()],
-                expected_all_drop_ops: vec!["s1".to_string()],
-                expected_user_drop_ops: vec![],
-                expected_all_create_ops: vec!["s1".to_string()],
-                expected_user_create_ops: vec![],
-                expected_migrated_system_object_mappings: vec!["s1".to_string()],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "child_migrations",
-                initial_state: vec![
-                    SimplifiedCatalogEntry {
-                        name: "s1".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "u1".to_string(),
-                        namespace: ItemNamespace::User,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s1".to_string()],
-                        },
-                    },
-                ],
-                migrated_names: vec!["s1".to_string()],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec!["u1".to_string()],
-                expected_previous_source_names: vec!["s1".to_string()],
-                expected_all_drop_ops: vec!["u1".to_string(), "s1".to_string()],
-                expected_user_drop_ops: vec!["u1".to_string()],
-                expected_all_create_ops: vec!["s1".to_string(), "u1".to_string()],
-                expected_user_create_ops: vec!["u1".to_string()],
-                expected_migrated_system_object_mappings: vec!["s1".to_string()],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "multi_child_migrations",
-                initial_state: vec![
-                    SimplifiedCatalogEntry {
-                        name: "s1".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "u1".to_string(),
-                        namespace: ItemNamespace::User,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s1".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "u2".to_string(),
-                        namespace: ItemNamespace::User,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s1".to_string()],
-                        },
-                    },
-                ],
-                migrated_names: vec!["s1".to_string()],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec!["u1".to_string(), "u2".to_string()],
-                expected_previous_source_names: vec!["s1".to_string()],
-                expected_all_drop_ops: vec!["u1".to_string(), "u2".to_string(), "s1".to_string()],
-                expected_user_drop_ops: vec!["u1".to_string(), "u2".to_string()],
-                expected_all_create_ops: vec!["s1".to_string(), "u2".to_string(), "u1".to_string()],
-                expected_user_create_ops: vec!["u2".to_string(), "u1".to_string()],
-                expected_migrated_system_object_mappings: vec!["s1".to_string()],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "topological_sort",
-                initial_state: vec![
-                    SimplifiedCatalogEntry {
-                        name: "s1".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s2".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "u1".to_string(),
-                        namespace: ItemNamespace::User,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s2".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "u2".to_string(),
-                        namespace: ItemNamespace::User,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s1".to_string(), "u1".to_string()],
-                        },
-                    },
-                ],
-                migrated_names: vec!["s1".to_string(), "s2".to_string()],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec!["u2".to_string(), "u1".to_string()],
-                expected_previous_source_names: vec!["s1".to_string(), "s2".to_string()],
-                expected_all_drop_ops: vec![
-                    "u2".to_string(),
-                    "s1".to_string(),
-                    "u1".to_string(),
-                    "s2".to_string(),
-                ],
-                expected_user_drop_ops: vec!["u2".to_string(), "u1".to_string()],
-                expected_all_create_ops: vec![
-                    "s2".to_string(),
-                    "u1".to_string(),
-                    "s1".to_string(),
-                    "u2".to_string(),
-                ],
-                expected_user_create_ops: vec!["u1".to_string(), "u2".to_string()],
-                expected_migrated_system_object_mappings: vec!["s1".to_string(), "s2".to_string()],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "topological_sort_complex",
-                initial_state: vec![
-                    SimplifiedCatalogEntry {
-                        name: "s273".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s322".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s317".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s349".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s273".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s421".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s273".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s295".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s273".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s296".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s295".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s320".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s295".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s340".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s295".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s318".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s295".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s323".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s295".to_string(), "s322".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s330".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s318".to_string(), "s317".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s321".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s318".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s315".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s296".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s354".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s296".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s327".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s296".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s339".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s296".to_string()],
-                        },
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s355".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::MaterializedView {
-                            referenced_names: vec!["s315".to_string()],
-                        },
-                    },
-                ],
-                migrated_names: vec![
-                    "s273".to_string(),
-                    "s317".to_string(),
-                    "s318".to_string(),
-                    "s320".to_string(),
-                    "s321".to_string(),
-                    "s322".to_string(),
-                    "s323".to_string(),
-                    "s330".to_string(),
-                    "s339".to_string(),
-                    "s340".to_string(),
-                ],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec![
-                    "s349".to_string(),
-                    "s421".to_string(),
-                    "s355".to_string(),
-                    "s315".to_string(),
-                    "s354".to_string(),
-                    "s327".to_string(),
-                    "s339".to_string(),
-                    "s296".to_string(),
-                    "s320".to_string(),
-                    "s340".to_string(),
-                    "s330".to_string(),
-                    "s321".to_string(),
-                    "s318".to_string(),
-                    "s323".to_string(),
-                    "s295".to_string(),
-                ],
-                expected_previous_source_names: vec![
-                    "s273".to_string(),
-                    "s317".to_string(),
-                    "s322".to_string(),
-                ],
-                expected_all_drop_ops: vec![
-                    "s349".to_string(),
-                    "s421".to_string(),
-                    "s355".to_string(),
-                    "s315".to_string(),
-                    "s354".to_string(),
-                    "s327".to_string(),
-                    "s339".to_string(),
-                    "s296".to_string(),
-                    "s320".to_string(),
-                    "s340".to_string(),
-                    "s330".to_string(),
-                    "s321".to_string(),
-                    "s318".to_string(),
-                    "s323".to_string(),
-                    "s295".to_string(),
-                    "s273".to_string(),
-                    "s317".to_string(),
-                    "s322".to_string(),
-                ],
-                expected_user_drop_ops: vec![],
-                expected_all_create_ops: vec![
-                    "s322".to_string(),
-                    "s317".to_string(),
-                    "s273".to_string(),
-                    "s295".to_string(),
-                    "s323".to_string(),
-                    "s318".to_string(),
-                    "s321".to_string(),
-                    "s330".to_string(),
-                    "s340".to_string(),
-                    "s320".to_string(),
-                    "s296".to_string(),
-                    "s339".to_string(),
-                    "s327".to_string(),
-                    "s354".to_string(),
-                    "s315".to_string(),
-                    "s355".to_string(),
-                    "s421".to_string(),
-                    "s349".to_string(),
-                ],
-                expected_user_create_ops: vec![],
-                expected_migrated_system_object_mappings: vec![
-                    "s322".to_string(),
-                    "s317".to_string(),
-                    "s273".to_string(),
-                    "s295".to_string(),
-                    "s323".to_string(),
-                    "s318".to_string(),
-                    "s321".to_string(),
-                    "s330".to_string(),
-                    "s340".to_string(),
-                    "s320".to_string(),
-                    "s296".to_string(),
-                    "s339".to_string(),
-                    "s327".to_string(),
-                    "s354".to_string(),
-                    "s315".to_string(),
-                    "s355".to_string(),
-                    "s421".to_string(),
-                    "s349".to_string(),
-                ],
-            },
-            BuiltinMigrationTestCase {
-                test_name: "system_child_migrations",
-                initial_state: vec![
-                    SimplifiedCatalogEntry {
-                        name: "s1".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Table,
-                    },
-                    SimplifiedCatalogEntry {
-                        name: "s2".to_string(),
-                        namespace: ItemNamespace::System,
-                        item: SimplifiedItem::Index {
-                            on: "s1".to_string(),
-                        },
-                    },
-                ],
-                migrated_names: vec!["s1".to_string()],
-                expected_previous_sink_names: vec![],
-                expected_previous_materialized_view_names: vec![],
-                expected_previous_source_names: vec!["s1".to_string()],
-                expected_all_drop_ops: vec!["s2".to_string(), "s1".to_string()],
-                expected_user_drop_ops: vec![],
-                expected_all_create_ops: vec!["s1".to_string(), "s2".to_string()],
-                expected_user_create_ops: vec![],
-                expected_migrated_system_object_mappings: vec!["s1".to_string(), "s2".to_string()],
-            },
-        ];
-
-        for test_case in test_cases {
-            Catalog::with_debug(NOW_ZERO.clone(), |mut catalog| async move {
-                let mut id_mapping = BTreeMap::new();
-                let mut name_mapping = BTreeMap::new();
-                for entry in test_case.initial_state {
-                    let (name, namespace, item) = entry.to_catalog_item(&id_mapping);
-                    let id = add_item(&mut catalog, name.clone(), item, namespace).await;
-                    id_mapping.insert(name.clone(), id);
-                    name_mapping.insert(id, name);
-                }
-
-                let migrated_ids = test_case
-                    .migrated_names
-                    .into_iter()
-                    .map(|name| id_mapping[&name])
-                    .collect();
-                let id_fingerprint_map: BTreeMap<GlobalId, String> = id_mapping
-                    .iter()
-                    .filter(|(_name, id)| id.is_system())
-                    // We don't use the new fingerprint in this test, so we can just hard code it
-                    .map(|(_name, id)| (*id, "".to_string()))
-                    .collect();
-                let migration_metadata = catalog
-                    .generate_builtin_migration_metadata(migrated_ids, id_fingerprint_map)
-                    .await
-                    .expect("failed to generate builtin migration metadata");
-
-                assert_eq!(
-                    convert_id_vec_to_name_vec(migration_metadata.previous_sink_ids, &name_mapping),
-                    test_case.expected_previous_sink_names,
-                    "{} test failed with wrong previous sink ids",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    convert_id_vec_to_name_vec(
-                        migration_metadata.previous_materialized_view_ids,
-                        &name_mapping
-                    ),
-                    test_case.expected_previous_materialized_view_names,
-                    "{} test failed with wrong previous materialized view ids",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    convert_id_vec_to_name_vec(
-                        migration_metadata.previous_source_ids,
-                        &name_mapping
-                    ),
-                    test_case.expected_previous_source_names,
-                    "{} test failed with wrong previous source ids",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    convert_id_vec_to_name_vec(migration_metadata.all_drop_ops, &name_mapping),
-                    test_case.expected_all_drop_ops,
-                    "{} test failed with wrong all drop ops",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    convert_id_vec_to_name_vec(migration_metadata.user_drop_ops, &name_mapping),
-                    test_case.expected_user_drop_ops,
-                    "{} test failed with wrong user drop ops",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    migration_metadata
-                        .all_create_ops
-                        .into_iter()
-                        .map(|(_, _, name, _, _, _)| name.item)
-                        .collect::<Vec<_>>(),
-                    test_case.expected_all_create_ops,
-                    "{} test failed with wrong all create ops",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    migration_metadata
-                        .user_create_ops
-                        .into_iter()
-                        .map(|(_, _, name)| name)
-                        .collect::<Vec<_>>(),
-                    test_case.expected_user_create_ops,
-                    "{} test failed with wrong user create ops",
-                    test_case.test_name
-                );
-                assert_eq!(
-                    migration_metadata
-                        .migrated_system_object_mappings
-                        .values()
-                        .map(|mapping| mapping.description.object_name.clone())
-                        .collect::<BTreeSet<_>>(),
-                    test_case
-                        .expected_migrated_system_object_mappings
-                        .into_iter()
-                        .collect::<BTreeSet<_>>(),
-                    "{} test failed with wrong migrated system object mappings",
-                    test_case.test_name
-                );
-            })
-            .await
-        }
     }
 
     #[mz_ore::test(tokio::test)]
